% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Action_QCCompletedIssues}
\alias{Action_QCCompletedIssues}
\title{Use a GitHub Action to QC completed issues}
\usage{
Action_QCCompletedIssues(lglOverwrite = FALSE, strPkgRoot = ".")
}
\arguments{
\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
The path to the created GitHub Action YAML file (invisibly).
}
\description{
Install a GitHub Action into a package repository to generate a QC report of
completed issues with \code{\link[=QCCompletedIssues]{QCCompletedIssues()}}. We recommend reviewing the
generated action to determine whether you would like to turn any features
off.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  Action_QCCompletedIssues()
\dontshow{\}) # examplesIf}
}
