% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Action_QCMilestone}
\alias{Action_QCMilestone}
\title{Use a GitHub Action to QC a milestone}
\usage{
Action_QCMilestone(lglOverwrite = FALSE, strPkgRoot = ".")
}
\arguments{
\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}
}
\value{
The path to the created GitHub Action YAML file (invisibly).
}
\description{
Install a GitHub Action into a package repository to generate a QC report
with \code{\link[=QCMilestones]{QCMilestones()}} of issues associated with a particular milestone. We
recommend reviewing the generated action to determine whether you would like
to turn any features off. Note: When triggered by a release, the workflow
looks for milestones that match the title of the release or the name of the
tag attached to the release. If the names do not match, the workflow will
fail.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  Action_QCPRIssues()
\dontshow{\}) # examplesIf}
}
