% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Action_QCPRIssues}
\alias{Action_QCPRIssues}
\title{Use a GitHub Action to QC pull-request-associated issues}
\usage{
Action_QCPRIssues(lglOverwrite = FALSE, strPkgRoot = ".")
}
\arguments{
\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
The path to the created GitHub Action YAML file (invisibly).
}
\description{
Install a GitHub Action into a package repository to generate a QC report
with \code{\link[=QCPR]{QCPR()}} of issues that will be closed by the triggering pull request.
We recommend reviewing the generated action to determine whether you would
like to turn any features off. Note: This workflow cannot be triggered when
an issue is connected to (or disconnected from) the pull request via the
"Development" UI of the PR or issue. In that situation, either trigger the
workflow manually, or edit the PR body to mention the issue (such as "Closes
#55").
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  Action_QCPRIssues()
\dontshow{\}) # examplesIf}
}
