% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Action_qcthat}
\alias{Action_qcthat}
\title{Use a GitHub Action to manage qcthat}
\usage{
Action_qcthat(lglOverwrite = FALSE, strPkgRoot = ".")
}
\arguments{
\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_find]{gert::git_find()}}.}
}
\value{
The path to the created GitHub Action YAML file (invisibly).
}
\description{
Install a GitHub Action into a package repository to manage qcthat Quality
Control with \code{\link[=TriggerUAT]{TriggerUAT()}}, \code{\link[=CommentAllReports]{CommentAllReports()}}, and
\code{\link[=AttachReleaseReports]{AttachReleaseReports()}}. We recommend reviewing the generated action to
determine whether you would like to turn any features off.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  Action_qcthat()
\dontshow{\}) # examplesIf}
}
