% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Action_UAT}
\alias{Action_UAT}
\title{Use a GitHub Action to manage UAT}
\usage{
Action_UAT(lglOverwrite = FALSE, strPkgRoot = ".")
}
\arguments{
\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
The path to the created GitHub Action YAML file (invisibly).
}
\description{
Install a GitHub Action into a package repository to manage the use
acceptance testing process with \code{\link[=TriggerUAT]{TriggerUAT()}}. The workflow triggers when
issues are closed, and, if they are labeled \code{"qcthat-uat"}, it triggers
reruns of the \code{\link[=Action_QCPRIssues]{Action_QCPRIssues()}} action for any open pull requests that
reference the UAT issue.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  Action_UAT()
\dontshow{\}) # examplesIf}
}
