% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{ApplyITMIgnoredLabels}
\alias{ApplyITMIgnoredLabels}
\title{Apply ignored labels to an Issue Test Matrix}
\usage{
ApplyITMIgnoredLabels(dfITM, chrIgnoredLabels)
}
\arguments{
\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}
}
\value{
A filtered \code{qcthat_IssueTestMatrix} object with issues that are
tagged to any of the specified ignored labels removed. The returned object
has an attribute \code{IgnoredLabels} which is a named list of of the issues
that were removed for each ignored label (as a \code{qcthat_IssueTestMatrix}),
or an empty named list if \code{chrIgnoredLabels} is empty.
}
\description{
Apply ignored labels to an Issue Test Matrix
}
\keyword{internal}
