% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{AsCommentsDF}
\alias{AsCommentsDF}
\title{Assign the qcthat_Comments class to a data frame}
\usage{
AsCommentsDF(x)
}
\arguments{
\item{x}{Object to assign the class to.}
}
\value{
A \code{qcthat_Comments} object.
}
\description{
Assign the qcthat_Comments class to a data frame
}
\keyword{internal}
