% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsExpectedDF}
\alias{AsExpectedDF}
\title{Assign a class and expected structure to a data frame}
\usage{
AsExpectedDF(df, dfShape, chrClass)
}
\arguments{
\item{df}{(\code{data.frame}) Data frame to assign the class to.}

\item{dfShape}{(\verb{0-row data.frame}) Data frame with the expected structure.}

\item{chrClass}{(\code{character}) Class name(s) to assign to the data frame.}
}
\value{
A data frame with the expected structure and class.
}
\description{
Assign a class and expected structure to a data frame
}
\keyword{internal}
