% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsExpectedFlat}
\alias{AsExpectedFlat}
\title{Listify a 1-row data frame and assign a class}
\usage{
AsExpectedFlat(x, lShape, chrClass)
}
\arguments{
\item{x}{(\code{data.frame} or \code{list}) Object to convert to a list.}

\item{lShape}{(\code{list}) List with the expected shape.}

\item{chrClass}{(\code{character}) Class name(s) to assign to the object.}
}
\value{
A list with the expected shape and class.
}
\description{
Listify a 1-row data frame and assign a class
}
\keyword{internal}
