% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{AsIssueTestResultsDF}
\alias{AsIssueTestResultsDF}
\title{Assign the qcthat_IssueTestResults class to a data frame}
\usage{
AsIssueTestResultsDF(df)
}
\arguments{
\item{df}{(\code{data.frame}) Data frame to assign the class to.}
}
\value{
A \code{qcthat_IssueTestResults} object.
}
\description{
Assign the qcthat_IssueTestResults class to a data frame
}
\keyword{internal}
