% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{AsMilestone}
\alias{AsMilestone}
\title{Listify a 1-row data frame and assign the qcthat_Milestone class}
\usage{
AsMilestone(x)
}
\arguments{
\item{x}{(\code{data.frame} or \code{list}) Object to convert to a list.}
}
\value{
A \code{qcthat_Milestone} object, which is a list with elements
\code{"Milestone"} and \code{"IssueTestResults"}.
}
\description{
Listify a 1-row data frame and assign the qcthat_Milestone class
}
\keyword{internal}
