% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{AsNestedIssueTestMatrix}
\alias{AsNestedIssueTestMatrix}
\title{Nest a qcthat_IssueTestMatrix by Milestone and Issue}
\usage{
AsNestedIssueTestMatrix(x)
}
\arguments{
\item{x}{(\code{qcthat_Object}) The qcthat object to format.}
}
\value{
A tibble with \code{TestResults} nested by \code{Issue}, and
\code{IssueTestResults} nested by \code{Milestone}.
}
\description{
Nest a qcthat_IssueTestMatrix by Milestone and Issue
}
\keyword{internal}
