% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{AsPRsDF}
\alias{AsPRsDF}
\title{Assign the qcthat_PRs class to a data frame}
\usage{
AsPRsDF(x)
}
\arguments{
\item{x}{Object to assign the class to.}
}
\value{
A \code{qcthat_PRs} object.
}
\description{
Assign the qcthat_PRs class to a data frame
}
\keyword{internal}
