% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsRowDFList}
\alias{AsRowDFList}
\title{Split a data frame by rows and transform}
\usage{
AsRowDFList(df, fnAsDF)
}
\arguments{
\item{df}{(\code{data.frame}) Data frame to split by rows.}

\item{fnAsDF}{(\code{function}) Function to apply to each single-row data frame.}
}
\value{
A list of objects, each processed via \code{fnAsDF}.
}
\description{
Split a data frame by rows and transform
}
\keyword{internal}
