% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignIssue.R
\name{AssignIssue}
\alias{AssignIssue}
\title{Assign a GitHub issue to one or more users}
\usage{
AssignIssue(
  dfIssue,
  chrAssignees = character(),
  lglOpenOnAssign = TRUE,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{dfIssue}{(\code{data.frame}, \code{numeric}, \code{gh_response}, or other) The issue
to assign, as returned by \code{\link[=FetchIssueDetails]{FetchIssueDetails()}}, or something that can be
coerced to such a \code{data.frame}.}

\item{chrAssignees}{(\code{character}) GitHub usernames to whom the issue should
be assigned (in addition to any current assignees). Elements will be split
on commas, so a single string can be read from an environment variable and
passed here.}

\item{lglOpenOnAssign}{(\verb{length-1 logical}) Whether to open the issue if it is
currently closed and at least one assignee is new.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A \code{qcthat_Issues} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{Title}: Issue title.
\item \code{Body}: Issue body (the full text of the issue).
\item \code{Labels}: List column of character vectors of issue labels.
\item \code{State}: Issue state (\code{open} or \code{closed}).
\item \code{StateReason}: Reason for issue state (e.g., \code{completed}) or \code{NA} if not
applicable.
\item \code{Milestone}: Issue milestone title or \code{NA} if not applicable.
\item \code{Type}: Issue type or \code{"Issue"} if no issue type is available.
\item \code{Url}: URL of the issue on GitHub.
\item \code{ParentOwner}: GitHub username or organization name of the parent issue
if applicable, otherwise \code{NA}.
\item \code{ParentRepo}: GitHub repository name of the parent issue if applicable,
otherwise \code{NA}.
\item \code{ParentNumber}: GitHub issue number of the parent issue if applicable,
otherwise \code{NA}.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the issue was created.
\item \code{ClosedAt}: \code{POSIXct} timestamp of when the issue was closed, or \code{NA} if
the issue is still open.
\item \code{Assignees}: List column of character vectors of GitHub usernames of
issue assignees.
}
}
\description{
Update the assignee list for a GitHub issue. If any members of \code{chrAssignees}
are not already assigned to the issue, they will be added as assignees, and
the issue will be opened if \code{lglOpenOnAssign} is \code{TRUE}.
}
