% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignIssue.R
\name{AssignIssueImpl}
\alias{AssignIssueImpl}
\title{Update a GitHub issue (implementation)}
\usage{
AssignIssueImpl(
  dfIssue,
  chrAssignees = character(),
  lglOpenOnAssign = TRUE,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{chrAssignees}{(\code{character}) GitHub usernames to which the issue
associated with an expectation should be assigned. Whenever the issue is
assigned to a new user, it will be re-opened. Elements of this vector will
be split on commas, so you can provide multiple assignees in a single
string. This is helpful if you would like to set up assignees via the
\code{"qcthat_UAT_ASSIGNEES"} environment variable, which is checked by default.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
A \code{qcthat_Issues} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{Title}: Issue title.
\item \code{Body}: Issue body (the full text of the issue).
\item \code{Labels}: List column of character vectors of issue labels.
\item \code{State}: Issue state (\code{open} or \code{closed}).
\item \code{StateReason}: Reason for issue state (e.g., \code{completed}) or \code{NA} if not
applicable.
\item \code{Milestone}: Issue milestone title or \code{NA} if not applicable.
\item \code{Type}: Issue type or \code{"Issue"} if no issue type is available.
\item \code{Url}: URL of the issue on GitHub.
\item \code{ParentOwner}: GitHub username or organization name of the parent issue
if applicable, otherwise \code{NA}.
\item \code{ParentRepo}: GitHub repository name of the parent issue if applicable,
otherwise \code{NA}.
\item \code{ParentNumber}: GitHub issue number of the parent issue if applicable,
otherwise \code{NA}.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the issue was created.
\item \code{ClosedAt}: \code{POSIXct} timestamp of when the issue was closed, or \code{NA} if
the issue is still open.
\item \code{Assignees}: List column of character vectors of GitHub usernames of
issue assignees.
}
}
\description{
Update a GitHub issue (implementation)
}
\keyword{internal}
