% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTestResults.R
\name{AsTestResultsDF}
\alias{AsTestResultsDF}
\title{Assign the qcthat_TestResults class to a data frame}
\usage{
AsTestResultsDF(x)
}
\arguments{
\item{x}{Object to assign the class to.}
}
\value{
A \code{qcthat_TestResults} object.
}
\description{
Assign the qcthat_TestResults class to a data frame
}
\keyword{internal}
