% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{BlameFile}
\alias{BlameFile}
\title{Git blame a file}
\usage{
BlameFile(strFilePath)
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{File}: The file name (character).
\item \code{Line}: Line number (integer).
\item \code{Commits}: List column where each element contains the commit SHA
(character) that last modified that line.
}
}
\description{
Git blame a file
}
\keyword{internal}
