% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{BlameFileRaw}
\alias{BlameFileRaw}
\title{Git blame a file (raw)}
\usage{
BlameFileRaw(strFilePath)
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}
}
\value{
The result of \code{\link[git2r:blame]{git2r::blame()}}. This function exists to make mocking
easier.
}
\description{
Git blame a file (raw)
}
\keyword{internal}
