% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{BlameFiles}
\alias{BlameFiles}
\title{Git blame multiple files}
\usage{
BlameFiles(chrFilePaths, envCall = rlang::caller_env())
}
\arguments{
\item{chrFilePaths}{(\code{character}) Paths to files to blame.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the same structure as \code{\link[=BlameFile]{BlameFile()}},
row-bound across all files.
}
\description{
Git blame multiple files
}
\keyword{internal}
