% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchUAIssue.R
\name{BodyUAIssue}
\alias{BodyUAIssue}
\title{Generate a body for a UAT issue}
\usage{
BodyUAIssue(chrChecks = character(), chrInstructions = character())
}
\arguments{
\item{chrChecks}{(\code{character}) Items for the user to check. These will be
preceded by checkboxes in the associated issue.}

\item{chrInstructions}{(\code{character}) Instructions for how to review an issue.
Included in the associated issue before the checklist.}
}
\value{
A string body for the UAT issue.
}
\description{
Generate a body for a UAT issue
}
\keyword{internal}
