% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllIssuePRRefs.R
\name{BuildIssuePRRefsQuery}
\alias{BuildIssuePRRefsQuery}
\title{Build a GraphQL query to find PRs referencing a specific issue}
\usage{
BuildIssuePRRefsQuery(intIssue)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}
}
\value{
A character string for the GraphQL query.
}
\description{
Build a GraphQL query to find PRs referencing a specific issue
}
\keyword{internal}
