% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllPRCommitSHAs.R
\name{BuildPRCommitsFragment}
\alias{BuildPRCommitsFragment}
\title{Build a GraphQL fragment for a single PR's commits}
\usage{
BuildPRCommitsFragment(intPR)
}
\arguments{
\item{intPR}{(\code{integer(1)}) PR number.}
}
\value{
A character string with the aliased GraphQL fragment.
}
\description{
Build a GraphQL fragment for a single PR's commits
}
\keyword{internal}
