% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllPRCommitSHAs.R
\name{BuildPRCommitsFragmentPaginated}
\alias{BuildPRCommitsFragmentPaginated}
\title{Build a paginated GraphQL fragment for a PR's commits}
\usage{
BuildPRCommitsFragmentPaginated(intPR, strCursor)
}
\arguments{
\item{intPR}{(\code{integer(1)}) PR number.}

\item{strCursor}{(\code{character(1)}) Pagination cursor.}
}
\value{
A character string with the GraphQL fragment.
}
\description{
Build a paginated GraphQL fragment for a PR's commits
}
\keyword{internal}
