% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{BuildPRIssuesQuery}
\alias{BuildPRIssuesQuery}
\title{Build a GraphQL sub-query for a single PR's issues}
\usage{
BuildPRIssuesQuery(intPRNumber)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to fetch
information about.}
}
\value{
A character string for the GraphQL sub-query.
}
\description{
Build a GraphQL sub-query for a single PR's issues
}
\keyword{internal}
