% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{CallGHAPI}
\alias{CallGHAPI}
\title{Wrapper around gh::gh() for mocking}
\usage{
CallGHAPI(
  strEndpoint,
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token(),
  ...
)
}
\arguments{
\item{strEndpoint}{(\verb{length-1 character}) The endpoint to call, e.g., \code{"GET /repos/{owner}/{repo}/issues"}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{...}{Additional parameters passed to \code{\link[gh:gh]{gh::gh()}}.}
}
\value{
The result of the \code{\link[gh:gh]{gh::gh()}} call.
}
\description{
Wrapper around gh::gh() for mocking
}
\keyword{internal}
