% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTestCode.R
\name{CheckFileLineOrder}
\alias{CheckFileLineOrder}
\title{Check that line numbers are in valid order}
\usage{
CheckFileLineOrder(intLineStart, intLineEnd, envCall = rlang::caller_env())
}
\arguments{
\item{intLineStart}{(\verb{length-1 integer}) The starting line number.}

\item{intLineEnd}{(\verb{length-1 integer}) The ending line number.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of validating
line order.
}
\description{
Check that line numbers are in valid order
}
\keyword{internal}
