% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{CheckPRIsOpen}
\alias{CheckPRIsOpen}
\title{Check if a pull request is open}
\usage{
CheckPRIsOpen(
  intPRNumber,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
\code{TRUE} if the pull request is open, \code{FALSE} otherwise.
}
\description{
Check if a pull request is open
}
\keyword{internal}
