% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckUAIssue.R
\name{CheckUAIssue}
\alias{CheckUAIssue}
\title{Check the status of a UAT issue}
\usage{
CheckUAIssue(
  strDescription,
  intIssue,
  chrInstructions = character(),
  chrChecks = character(),
  lglReportFailure = IsCheckingUAT(),
  chrAssignees = Sys.getenv("qcthat_UAT_ASSIGNEES"),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strDescription}{(\verb{length-1 character}) A brief description of a user
expectation.}

\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{chrInstructions}{(\code{character}) Instructions for how to review an issue.
Included in the associated issue before the checklist.}

\item{chrChecks}{(\code{character}) Items for the user to check. These will be
preceded by checkboxes in the associated issue.}

\item{lglReportFailure}{(\verb{length-1 logical}) Whether to ignore failures
(default unless a "qcthat_UAT" environment variable is "true"), or fail
(and show as a failure in testthat tests).}

\item{chrAssignees}{(\code{character}) GitHub usernames to which the issue
associated with an expectation should be assigned. Whenever the issue is
assigned to a new user, it will be re-opened. Elements of this vector will
be split on commas, so you can provide multiple assignees in a single
string. This is helpful if you would like to set up assignees via the
\code{"qcthat_UAT_ASSIGNEES"} environment variable, which is checked by default.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
The value of \code{envQcthat$UATIssues}, invisibly (called for side
effects).
}
\description{
Check the status of a UAT issue
}
\keyword{internal}
