% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{ChooseCoverageFooter}
\alias{ChooseCoverageFooter}
\title{Choose coverage footer}
\usage{
ChooseCoverageFooter(
  intIssues,
  chrTests,
  lglUseEmoji = getOption("qcthat.emoji", TRUE)
)
}
\arguments{
\item{intIssues}{(\code{integer}) A vector of issue numbers from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix or from GitHub.}

\item{chrTests}{(\code{character}) A vector of test descriptions from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A string with or without an emoji describing test coverage of
issues, or \code{NULL} if there aren't any issues.
}
\description{
Choose coverage footer
}
\keyword{internal}
