% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSingleIssueTestResults.R
\name{ChooseDispositionIndicator}
\alias{ChooseDispositionIndicator}
\title{Choose emoji to indicate test disposition}
\usage{
ChooseDispositionIndicator(
  chrDisposition,
  lglUseEmoji = getOption("qcthat.emoji", TRUE)
)
}
\arguments{
\item{chrDisposition}{(\code{character}) Test disposition. Generally one of
\code{"pass"}, \code{"fail"}, or \code{"skip"}.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A character vector of the same length as \code{chrDisposition}, with each
element being the emoji or ASCII indicator for the corresponding test
disposition.
}
\description{
Choose emoji to indicate test disposition
}
\keyword{internal}
