% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{ChooseEmoji}
\alias{ChooseEmoji}
\title{Helper to choose emojis or ASCII indicators}
\usage{
ChooseEmoji(strCondition, lglUseEmoji = getOption("qcthat.emoji", TRUE))
}
\arguments{
\item{strCondition}{(\verb{length-1 character}) The condition to create a key item
for.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A length-1 character vector representing the emoji or ASCII
indicator for the given condition.
}
\description{
Helper to choose emojis or ASCII indicators
}
\keyword{internal}
