% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{ChooseOverallDispositionMessage}
\alias{ChooseOverallDispositionMessage}
\title{Choose overall disposition message}
\usage{
ChooseOverallDispositionMessage(fctDisposition)
}
\arguments{
\item{fctDisposition}{(\code{factor}) Disposition factor with levels \code{c("fail", "skip", "pass")}.}
}
\value{
A string summary message or \code{NULL} if no disposition is available.
}
\description{
Choose overall disposition message
}
\keyword{internal}
