% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{ChooseRefPRNumber}
\alias{ChooseRefPRNumber}
\title{Choose between multiple PRs}
\usage{
ChooseRefPRNumber(dfPRs, strSourceRef = GetActiveBranch())
}
\arguments{
\item{dfPRs}{(\code{data.frame}) Data frame of pull requests as returned by
\code{\link[=FetchRepoPRs]{FetchRepoPRs()}}.}

\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}
}
\value{
A \verb{length-1 integer} pull request number.
}
\description{
Choose between multiple PRs
}
\keyword{internal}
