% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{ChooseTreeCharacter}
\alias{ChooseTreeCharacter}
\title{Helper to choose tree characters}
\usage{
ChooseTreeCharacter(strCharacterDescription)
}
\arguments{
\item{strCharacterDescription}{(\verb{length-1 character}) The special character
to choose.}
}
\value{
A length-1 character vector representing the chosen tree character.
}
\description{
R CMD check doesn't like these characters appearing "raw", and they're hard
to read as U codes, so encode them in one place.
}
\keyword{internal}
