% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateUAIssue.R
\name{ClearGHCache}
\alias{ClearGHCache}
\title{Clear the cache used by gh}
\usage{
ClearGHCache()
}
\value{
\code{0} for success, \code{1} for failure, invisibly (as returned by
\code{\link[=unlink]{unlink()}}).
}
\description{
Clear the cache used by gh
}
\keyword{internal}
