% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CollapseGluedData}
\alias{CollapseGluedData}
\title{Convenience function to glue data and collapse with a separator}
\usage{
CollapseGluedData(dfData, ..., strSep = "\\n")
}
\arguments{
\item{dfData}{(\code{data.frame} or \code{list}) Data frame or list to use as the data
source for \code{\link[glue:glue_data]{glue::glue_data()}}.}

\item{...}{Expressions to evaluate within the context of \code{dfData}, passed to
\code{\link[glue:glue_data]{glue::glue_data()}}.}

\item{strSep}{(\verb{length-1 character}) Separator to use between glued values.
Defaults to a newline, which is common for formatting multiple lines of
text in GitHub comments.}
}
\value{
A string containing the glued and collapsed result.
}
\description{
Convenience function to glue data and collapse with a separator
}
\keyword{internal}
