% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentAllReports.R
\name{CommentAllReports}
\alias{CommentAllReports}
\title{Comment on a PR or issue with QC reports}
\usage{
CommentAllReports(
  intPRNumber = GuessPRNumber(strOwner = strOwner, strRepo = strRepo, strGHToken =
    strGHToken),
  lglCompleted = TRUE,
  lglMilestone = length(chrMilestones),
  lglPR = TRUE,
  lglUAT = TRUE,
  chrMilestones = GuessMilestones(),
  dfITM = NULL,
  lglUpdate = TRUE,
  strRunID = Sys.getenv("GITHUB_RUN_ID"),
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{lglCompleted}{(\verb{length-1 logical}) Whether to include the
\code{\link[=QCCompletedIssues]{QCCompletedIssues()}} report.}

\item{lglMilestone}{(\verb{length-1 logical}) Whether to include the
\code{\link[=QCMilestones]{QCMilestones()}} report.}

\item{lglPR}{(\verb{length-1 logical}) Whether to include the \code{\link[=QCPR]{QCPR()}}
report.}

\item{lglUAT}{(\verb{length-1 logical}) Whether to include the \code{\link[=CommentUAT]{CommentUAT()}}
report.}

\item{chrMilestones}{(\code{character}) The name(s) of milestone(s) to filter
issues by.}

\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}} (often via \code{\link[=QCPackage]{QCPackage()}}).}

\item{lglUpdate}{(\verb{length-1 logical}) Whether to update an existing comment
if it already exists (rather than creating a new comment).}

\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
\code{dfITM}, invisibly.
}
\description{
Add or update a comment on a GitHub pull request (or issue) with one or more
QC reports, formatted in GitHub markdown.
}
