% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentUAT.R
\name{CommentUAT}
\alias{CommentUAT}
\title{Comment on a PR or issue with a UAT report}
\usage{
CommentUAT(
  intPRNumber = GuessPRNumber(strOwner = strOwner, strRepo = strRepo, strGHToken =
    strGHToken),
  lglUpdate = TRUE,
  strRunID = Sys.getenv("GITHUB_RUN_ID"),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{lglUpdate}{(\verb{length-1 logical}) Whether to update an existing comment
or label if it already exists (rather than creating a new comment or
label).}

\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
Invisibly returns the result of \code{\link[=CommentIssue]{CommentIssue()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Add or update a comment on a GitHub pull request (or issue) with a report of
issues that require user acceptance, formatted in GitHub markdown. Note: This
should only be called \emph{after} the test suite has ran.
}
