% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{CompileCommentsDF}
\alias{CompileCommentsDF}
\title{Compile comments data frame}
\usage{
CompileCommentsDF(lCommentsRaw)
}
\arguments{
\item{lCommentsRaw}{(\code{list}) List of raw comment objects as returned by
\code{\link[gh:gh]{gh::gh()}}.}
}
\value{
A \code{qcthat_Comments} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{Body}: Comment body (the full text of the comment).
\item \code{Url}: URL of the comment on GitHub.
\item \code{CommentGHID}: GitHub ID of the comment.
\item \code{Author}: GitHub username of the comment author.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the issue was created.
\item \code{UpdatedAt}: \code{POSIXct} timestamp of when the issue was last updated.
\item \code{qcthatCommentID}: The \code{qcthat-comment-id}, which is usually a hash of
the title of the comment (when present).
}
}
\description{
Compile comments data frame
}
\keyword{internal}
