% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{CompileConditionClasses}
\alias{CompileConditionClasses}
\title{Paste together standardized pieces of a condition class}
\usage{
CompileConditionClasses(
  strConditionSubclass,
  strConditionClass = c("error", "warning", "message")
)
}
\arguments{
\item{strConditionSubclass}{(\verb{length-1 character}) A subclass for a
condition.}

\item{strConditionClass}{(\verb{length-1 character}) One of "error", "warning", or
"message".}
}
\value{
A character vector of condition classes.
}
\description{
Paste together standardized pieces of a condition class
}
\keyword{internal}
