% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTestResults.R
\name{CompileDispositions}
\alias{CompileDispositions}
\title{Extract disposition information from testthat results}
\usage{
CompileDispositions(lTestResults)
}
\arguments{
\item{lTestResults}{(\code{testthat_results}) A testthat test results object,
typically obtained by running something like \code{\link[testthat:test_package]{testthat::test_local()}} with
\code{stop_on_failure = FALSE} and a reporter that doesn't cause issues in
parallel testing, like \code{reporter = "silent"}, and assigning it to a name.}
}
\value{
A factor with levels \code{pass}, \code{fail}, and \code{skip}.
}
\description{
Extract disposition information from testthat results
}
\keyword{internal}
