% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{CompileGitLogDiff}
\alias{CompileGitLogDiff}
\title{Create a data frame of commits that are in one git ref but not another}
\usage{
CompileGitLogDiff(
  strSourceRef,
  strTargetRef,
  strPkgRoot = ".",
  intMaxCommits = 100000L
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{intMaxCommits}{(\verb{length-1 integer}) The maximum number of commits to
return from git logs. Leaving this at the default should almost always be
fine, but you can reduce the number if your repository has a long commit
history and this function is slow.}
}
\value{
A data frame of commits that are in \code{strSourceRef} but not in
\code{strTargetRef}, as returned by \code{\link[=GetGitLog]{GetGitLog()}}.
}
\description{
Create a data frame of commits that are in one git ref but not another
}
\keyword{internal}
