% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{CompileIssueTestMatrix}
\alias{CompileIssueTestMatrix}
\title{Create a nested tibble of issues and tests}
\usage{
CompileIssueTestMatrix(dfRepoIssues, dfTestResults)
}
\arguments{
\item{dfRepoIssues}{(\code{qcthat_Issues} or data frame) Data frame of GitHub
issues as returned by \code{\link[=FetchRepoIssues]{FetchRepoIssues()}}.}

\item{dfTestResults}{(\code{qcthat_TestResults} or data frame) Data frame of test
results as returned by \code{\link[=CompileTestResults]{CompileTestResults()}}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object, which is a nested tibble with
columns:
\itemize{
\item \code{Milestone}: The milestone title associated with the issues.
\item \code{IssueTestResults}: A nested tibble with the other columns returned by
\code{\link[=FetchRepoIssues]{FetchRepoIssues()}}, plus \code{TestResults}, which is a nested tibble with
the remaining columns from \code{\link[=CompileTestResults]{CompileTestResults()}}.
}
}
\description{
Combine the data from \code{\link[=CompileTestResults]{CompileTestResults()}} and \code{\link[=FetchRepoIssues]{FetchRepoIssues()}} into a
nested tibble organized by milestone, with each milestone containing issues
and associated tests.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
  lTestResults <- testthat::test_local(stop_on_failure = FALSE)
  CompileIssueTestMatrix(
    dfRepoIssues = FetchRepoIssues(),
    dfTestResults = CompileTestResults(lTestResults)
  )
\dontshow{\}) # examplesIf}
}
