% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{CompileIssueTestResultsByMilestone}
\alias{CompileIssueTestResultsByMilestone}
\title{Compile issue test results by milestone}
\usage{
CompileIssueTestResultsByMilestone(dfRepoIssues, dfTestResultsByIssue)
}
\arguments{
\item{dfRepoIssues}{(\code{qcthat_Issues} or data frame) Data frame of GitHub
issues as returned by \code{\link[=FetchRepoIssues]{FetchRepoIssues()}}.}

\item{dfTestResultsByIssue}{(\code{data.frame}) Data frame with \code{TestResults}
nested by \code{Issue}, as returned by \code{\link[=CompileTestResultsByIssue]{CompileTestResultsByIssue()}}.}
}
\value{
A data frame with \code{IssueTestResults} nested by \code{Milestone}
}
\description{
Compile issue test results by milestone
}
\keyword{internal}
