% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttachReleaseReports.R
\name{CompileReleaseReportsMarkdown}
\alias{CompileReleaseReportsMarkdown}
\title{Compile release reports in GitHub markdown}
\usage{
CompileReleaseReportsMarkdown(
  chrBody,
  strRunID = Sys.getenv("GITHUB_RUN_ID"),
  strJobName = Sys.getenv("GITHUB_JOB"),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strJobName}{(\verb{length-1 character}) Name of a GitHub Actions workflow
job.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A single string containing the compiled markdown.
}
\description{
Compile release reports in GitHub markdown
}
\keyword{internal}
