% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTestResults.R
\name{CompileTestResults}
\alias{CompileTestResults}
\title{Extract information from testthat results}
\usage{
CompileTestResults(lTestResults)
}
\arguments{
\item{lTestResults}{A \code{testthat_results} object, typically obtained by
running something like \code{testthat::test_local(stop_on_failure = FALSE)} and
assigning it to a name.}
}
\value{
A tibble with columns:
\itemize{
\item \code{test}: Name of the test.
\item \code{file}: File where the test is located.
\item \code{disposition}: Factor with levels \code{pass}, \code{fail}, and \code{skip} indicating the
overall outcome of the test.
\item \code{issues}: List column containing integer vectors of associated GitHub issue
numbers extracted from the test names.
}
}
\description{
Extract relevant information from a \code{testthat_results} object into a tidy
tibble format.
}
\examples{
# Generate a test results object.

# lTestResults <- testthat::test_local(stop_on_failure = FALSE)

lTestResults <- structure(
  list(
    list(
      file = "test-file1.R",
      test = "First test with one GH issue (#32)",
      results = list(
        structure(
          list(),
          class = c("expectation_success", "expectation", "condition")
        )
      )
    ),
    list(
      file = "test-file1.R",
      test = "Second test with two GH issues (#32, #45)",
      results = list(
        structure(
          list(),
          class = c("expectation_failure", "expectation", "error", "condition")
        )
      )
    )
  ),
  class = "testthat_results"
)
CompileTestResults(lTestResults)
}
