% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{CompileTestResultsByIssue}
\alias{CompileTestResultsByIssue}
\title{Compile test results by issue}
\usage{
CompileTestResultsByIssue(dfTestResults)
}
\arguments{
\item{dfTestResults}{(\code{qcthat_TestResults} or data frame) Data frame of test
results as returned by \code{\link[=CompileTestResults]{CompileTestResults()}}.}
}
\value{
A data frame with \code{TestResults} nested by \code{Issue}
}
\description{
Compile test results by issue
}
\keyword{internal}
