% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CompletelyFlatten}
\alias{CompletelyFlatten}
\title{Get rid of list structure}
\usage{
CompletelyFlatten(lPuffy)
}
\arguments{
\item{lPuffy}{(\code{list}) A potentially nested list with one type of object you
want to keep.}
}
\value{
A sorted, unnamed vector of unique values from the unlisted list.
}
\description{
Get rid of list structure
}
\keyword{internal}
