% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateUAIssue.R
\name{ConnectChildIssueByID}
\alias{ConnectChildIssueByID}
\title{Connect an issue to a parent issue}
\usage{
ConnectChildIssueByID(
  strChildIssueID,
  intParentIssue,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strChildIssueID}{(\verb{length-1 character}) The \code{id} field of an issue to
connect to a parent issue.}

\item{intParentIssue}{(\verb{length-1 integer}) The number of the parent issue in
a parent-child issue relationship.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The URL of the parent issue.
}
\description{
Connect an issue to a parent issue
}
\keyword{internal}
