% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{ConstructJobURL}
\alias{ConstructJobURL}
\title{Construct the URL for a GitHub Actions job}
\usage{
ConstructJobURL(
  strRunID = character(),
  strJobID = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strJobID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run job.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}
}
\value{
A string containing the URL for the specified GitHub Actions job.
}
\description{
Construct the URL for a GitHub Actions job
}
\keyword{internal}
