% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentUAT.R
\name{ConstructUATBody}
\alias{ConstructUATBody}
\title{Construct the body of the UAT report comment}
\usage{
ConstructUATBody(dfRelevantUAT)
}
\arguments{
\item{dfRelevantUAT}{(\code{data.frame}) A data frame with columns \code{ParentIssue},
\code{UATIssue}, \code{Description}, \code{Disposition}, \code{Owner}, and \code{Repo}.}
}
\value{
A string containing the body of the UAT report comment, formatted in
GitHub markdown.
}
\description{
Construct the body of the UAT report comment
}
\keyword{internal}
