% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentUAT.R
\name{ConstructUATBodyPiece}
\alias{ConstructUATBodyPiece}
\title{Construct the body of the UAT report comment}
\usage{
ConstructUATBodyPiece(dfIssues, strHeader, strBulletTemplate, strFooter = NULL)
}
\arguments{
\item{dfIssues}{(\code{data.frame}) A data frame with columns \code{ParentIssue},
\code{UATIssue}, \code{Description}, \code{Disposition}, \code{Owner}, and \code{Repo}.}

\item{strHeader}{(\verb{length-1 character}) A header to introduce the issues in
this section of the report.}

\item{strBulletTemplate}{(\verb{length-1 character}) A template for the bullet
points describing each issue in this section of the report, with glue
syntax for the relevant columns of \code{dfIssues}.}

\item{strFooter}{(\verb{length-1 character} or \code{NULL}) An optional footer to
conclude this section of the report.}
}
\value{
A string containing the body of this part of the UAT report comment,
formatted in GitHub markdown.
}
\description{
Construct the body of the UAT report comment
}
\keyword{internal}
