% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CountNonNA}
\alias{CountNonNA}
\title{Convenience function to count non-NA unique values}
\usage{
CountNonNA(x)
}
\arguments{
\item{x}{(\code{vector}) Vector to count unique non-NA values in.}
}
\value{
An integer representing the number of unique non-NA values in \code{x}.
}
\description{
Convenience function to count non-NA unique values
}
\keyword{internal}
