% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{CoverSingleTestRaw}
\alias{CoverSingleTestRaw}
\title{Run covr::environment_coverage (raw wrapper for mocking)}
\usage{
CoverSingleTestRaw(envPkg, strTempFile)
}
\arguments{
\item{envPkg}{(\code{environment}) Loaded package environment.}

\item{strTempFile}{(\verb{length-1 character}) Path to a temporary test file.}
}
\value{
A \code{coverage} object.
}
\description{
Run covr::environment_coverage (raw wrapper for mocking)
}
\keyword{internal}
