% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateChildIssue.R
\name{CreateChildIssue}
\alias{CreateChildIssue}
\title{Create a child issue linked to a parent issue}
\usage{
CreateChildIssue(
  intParentIssue,
  strTitle,
  strBody,
  ...,
  chrLabels = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intParentIssue}{(\verb{length-1 integer}) The number of an issue to which a
child issue will attach.}

\item{strTitle}{(\verb{length-1 character}) A title for an issue.}

\item{strBody}{(\verb{length-1 character}) The body of an issue, in GitHub
markdown.}

\item{...}{Additional parameters passed to \code{\link[=CreateRepoIssueRaw]{CreateRepoIssueRaw()}}.}

\item{chrLabels}{(\code{character}) The name(s) of labels(s) to use.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A tibble with one row representing the created child issue.
}
\description{
Create a child issue linked to a parent issue
}
\keyword{internal}
