% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGHLabel.R
\name{CreateGHLabelImpl}
\alias{CreateGHLabelImpl}
\title{Create a GitHub label (implementation)}
\usage{
CreateGHLabelImpl(
  strLabel,
  strLabelDescription = "{qcthat}: A new label",
  strLabelColor = "#444444",
  lglVerbose = getOption("qcthat-verbose", FALSE),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strLabel}{(\verb{length-1 character}) The name of the label to create or
update.}

\item{strLabelDescription}{(\verb{length-1 character}) The description for the
label.}

\item{strLabelColor}{(\verb{length-1 character}) The hex color code for the label
(e.g., \code{"#444444"}).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
The raw label object as returned by \code{\link[gh:gh]{gh::gh()}} (invisibly).
}
\description{
Create a GitHub label (implementation)
}
\keyword{internal}
