% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateChildIssue.R
\name{CreateRepoIssueRaw}
\alias{CreateRepoIssueRaw}
\title{Create an issue in a repository}
\usage{
CreateRepoIssueRaw(
  strTitle,
  strBody,
  ...,
  chrLabels = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strTitle}{(\verb{length-1 character}) A title for an issue.}

\item{strBody}{(\verb{length-1 character}) The body of an issue, in GitHub
markdown.}

\item{...}{Additional parameters passed to \code{\link[=CallGHAPI]{CallGHAPI()}}.}

\item{chrLabels}{(\code{character}) The name(s) of labels(s) to use.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A list representing the issue, as returned by GitHub.
}
\description{
Create an issue in a repository
}
\keyword{internal}
