% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{DefaultIgnoreLabelDescriptions}
\alias{DefaultIgnoreLabelDescriptions}
\title{Default descriptions for ignored labels}
\usage{
DefaultIgnoreLabelDescriptions()
}
\value{
A character vector of label descriptions.
}
\description{
Returns the character vector of descriptions corresponding to the labels
returned by \code{\link[=DefaultIgnoreLabels]{DefaultIgnoreLabels()}}.
}
\keyword{internal}
