% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{DefaultIgnoreLabels}
\alias{DefaultIgnoreLabels}
\title{Default labels to ignore}
\usage{
DefaultIgnoreLabels()
}
\value{
A character vector of label names.
}
\description{
Returns the character vector of issue labels that are ignored by default in
QC reports. Currently, this list only includes \code{"qcthat-nocov"} and
\code{"qcthat-uat"}, but it may change as we add more standard labels.
}
\examples{
DefaultIgnoreLabels()
}
