% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{DefaultIgnoreLabelsDF}
\alias{DefaultIgnoreLabelsDF}
\title{Default ignored labels as a tibble}
\usage{
DefaultIgnoreLabelsDF()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{Label}, \code{Description}, and
\code{Color}.
}
\description{
Returns a tibble of ignore labels (from \code{\link[=DefaultIgnoreLabels]{DefaultIgnoreLabels()}}), their
descriptions, and the colors of their labels.
}
\examples{
DefaultIgnoreLabelsDF()
}
