% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTestResults.R
\name{.CompileDispositions}
\alias{.CompileDispositions}
\title{Extract disposition information from testthat results}
\usage{
.CompileDispositions(lTestResults)
}
\arguments{
\item{lTestResults}{A \code{testthat_results} object, typically obtained by
running something like \code{testthat::test_local(stop_on_failure = FALSE)} and
assigning it to a name.}
}
\value{
A factor with levels \code{pass}, \code{fail}, and \code{skip}.
}
\description{
Extract disposition information from testthat results
}
\keyword{internal}
