% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{EmptyCommentsDF}
\alias{EmptyCommentsDF}
\title{Empty comments data frame}
\usage{
EmptyCommentsDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty comments data frame
}
\keyword{internal}
