% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{EmptyCommentsDFRaw}
\alias{EmptyCommentsDFRaw}
\title{Empty comments data frame (raw)}
\usage{
EmptyCommentsDFRaw()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns (before
transformations) but no rows.
}
\description{
Empty comments data frame (raw)
}
\keyword{internal}
