% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTestsFromFiles.R
\name{EmptyFileTestsDF}
\alias{EmptyFileTestsDF}
\title{Empty file tests data frame}
\usage{
EmptyFileTestsDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty file tests data frame
}
\keyword{internal}
