% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{EmptyGHResponse}
\alias{EmptyGHResponse}
\title{A classed GitHub response for testing}
\usage{
EmptyGHResponse()
}
\value{
An empty list with class "gh_response".
}
\description{
A classed GitHub response for testing
}
\keyword{internal}
