% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueDetails.R
\name{EmptyIssueDetailsDF}
\alias{EmptyIssueDetailsDF}
\title{Create empty issue details data frame}
\usage{
EmptyIssueDetailsDF()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{Title}: Issue title.
\item \code{Body}: Issue body (or \code{NA_character_} if NULL).
}
}
\description{
Create empty issue details data frame
}
\keyword{internal}
