% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{EmptyIssuesDF}
\alias{EmptyIssuesDF}
\title{Empty issues data frame}
\usage{
EmptyIssuesDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty issues data frame
}
\keyword{internal}
