% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{EmptyIssuesDFRaw}
\alias{EmptyIssuesDFRaw}
\title{Empty issues data frame (raw)}
\usage{
EmptyIssuesDFRaw()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns (before
transformations) but no rows.
}
\description{
Empty issues data frame (raw)
}
\keyword{internal}
