% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{EmptyIssueTestResultsDF}
\alias{EmptyIssueTestResultsDF}
\title{Empty issue test results data frame}
\usage{
EmptyIssueTestResultsDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty issue test results data frame
}
\keyword{internal}
