% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGHLabels.R
\name{EmptyLabelsDF}
\alias{EmptyLabelsDF}
\title{Create an empty labels data frame}
\usage{
EmptyLabelsDF()
}
\value{
A 0-row \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{Label}, \code{Description}, and
\code{Color}.
}
\description{
Create an empty labels data frame
}
\keyword{internal}
