% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{EmptyPRsDF}
\alias{EmptyPRsDF}
\title{Empty PR data frame}
\usage{
EmptyPRsDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty PR data frame
}
\keyword{internal}
