% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{EmptyPRsDFRaw}
\alias{EmptyPRsDFRaw}
\title{Empty PR data frame (raw)}
\usage{
EmptyPRsDFRaw()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns (before
transformations) but no rows.
}
\description{
Empty PR data frame (raw)
}
\keyword{internal}
