% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{EmptyTestCoveredLinesDF}
\alias{EmptyTestCoveredLinesDF}
\title{Empty test covered lines data frame}
\usage{
EmptyTestCoveredLinesDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty test covered lines data frame
}
\keyword{internal}
