% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTestIssueContext.R
\name{EmptyTestIssueContextDF}
\alias{EmptyTestIssueContextDF}
\title{Create empty test issue context data frame}
\usage{
EmptyTestIssueContextDF()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Test}: The \code{desc} field of the test from \code{\link[testthat:test_that]{testthat::test_that()}}.
\item \code{File}: Path to the file where the test is defined, relative to the
package root.
\item \code{LineStart}: Starting line number of test.
\item \code{LineEnd}: Ending line number of test.
\item \code{Issues}: List column containing integer vectors of issue numbers already
tagged in the test description.
\item \code{PotentialIssueDetails}: List column of tibbles with enriched issue
details.
\item \code{TestCode}: List column of character vectors containing test code.
}
}
\description{
Create empty test issue context data frame
}
\keyword{internal}
