% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestFilesToPotentialIssues.R
\name{EmptyTestPotentialIssues}
\alias{EmptyTestPotentialIssues}
\title{Empty test potential issues data frame}
\usage{
EmptyTestPotentialIssues()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty test potential issues data frame
}
\keyword{internal}
