% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTestResults.R
\name{EmptyTestResultsDF}
\alias{EmptyTestResultsDF}
\title{Empty test results data frame}
\usage{
EmptyTestResultsDF()
}
\value{
A standard \code{\link[tibble:tibble]{tibble::tibble()}} with the correct columns but no rows.
}
\description{
Empty test results data frame
}
\keyword{internal}
