% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{EnframeComments}
\alias{EnframeComments}
\title{Transform comments list into tibble with expected columns}
\usage{
EnframeComments(lCommentsRaw)
}
\arguments{
\item{lCommentsRaw}{(\code{list}) List of raw comment objects as returned by
\code{\link[gh:gh]{gh::gh()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with raw comments data.
}
\description{
Transform comments list into tibble with expected columns
}
\keyword{internal}
