% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGHLabels.R
\name{EnframeGHLabels}
\alias{EnframeGHLabels}
\title{Convert raw GitHub labels list to data frame}
\usage{
EnframeGHLabels(lGHLabels)
}
\arguments{
\item{lGHLabels}{(\code{list}) List of label objects as returned by \code{\link[gh:gh]{gh::gh()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{Label}, \code{Description}, and
\code{Color}, or \code{NULL} if \code{lGHLabels} is empty.
}
\description{
Convert raw GitHub labels list to data frame
}
\keyword{internal}
