% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{EnframeHunk}
\alias{EnframeHunk}
\title{Turn a hunk list into a tibble}
\usage{
EnframeHunk(lHunk, strFilePath)
}
\arguments{
\item{lHunk}{(\code{list}) A single element of \code{lBlameRaw$hunks} as returned by
\code{\link[=BlameFileRaw]{BlameFileRaw()}}.}

\item{strFilePath}{(\verb{length-1 character}) A file path.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{File}: The file name (character).
\item \code{Line}: Line number (integer).
\item \code{Commits}: List column where each element contains the commit SHA
(character) that last modified that line.
}
}
\description{
Turn a hunk list into a tibble
}
\keyword{internal}
