% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{EnframeIssues}
\alias{EnframeIssues}
\title{Transform issues list into tibble with expected columns}
\usage{
EnframeIssues(lIssuesNonPR)
}
\arguments{
\item{lIssuesNonPR}{(\code{list}) List of issue objects as returned by
\code{\link[=RemovePRsFromIssues]{RemovePRsFromIssues()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with raw issue data.
}
\description{
Transform issues list into tibble with expected columns
}
\keyword{internal}
