% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{EnframePRs}
\alias{EnframePRs}
\title{Transform PR list into tibble with expected columns}
\usage{
EnframePRs(lPRs)
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with raw PR data.
}
\description{
Transform PR list into tibble with expected columns
}
\keyword{internal}
