% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTestIssueContext.R
\name{EnrichWithIssueDetails}
\alias{EnrichWithIssueDetails}
\title{Enrich data frame with issue details}
\usage{
EnrichWithIssueDetails(dfPotentialIssues, strOwner, strRepo, strGHToken)
}
\arguments{
\item{dfPotentialIssues}{(\code{tibble}) A data frame as returned by
\code{\link[=MapTestFilesToPotentialIssues]{MapTestFilesToPotentialIssues()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
The input data frame with \code{PotentialIssues} enriched to
\code{PotentialIssueDetails} with issue details.
}
\description{
Enrich data frame with issue details
}
\keyword{internal}
