% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTestIssueContext.R
\name{EnrichWithTestCode}
\alias{EnrichWithTestCode}
\title{Enrich data frame with test code}
\usage{
EnrichWithTestCode(dfTestPotentialIssueDetails, strTestDir)
}
\arguments{
\item{dfTestPotentialIssueDetails}{(\code{tibble}) A data frame as returned by
\code{\link[=EnrichWithIssueDetails]{EnrichWithIssueDetails()}}.}

\item{strTestDir}{(\verb{length-1 character}) Path to the directory containing
test files. Defaults to \code{"tests/testthat"}.}
}
\value{
The input data frame with a \code{TestCode} list column added.
}
\description{
Enrich data frame with test code
}
\keyword{internal}
