% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{ExpandLines}
\alias{ExpandLines}
\title{Stretch LineStart and LineEnd to Line rows}
\usage{
ExpandLines(dfFileTests)
}
\arguments{
\item{dfFileTests}{(\code{data.frame}) A \code{\link[tibble:tibble]{tibble::tibble()}} with the structure
returned by \code{\link[=ExtractTestsFromFiles]{ExtractTestsFromFiles()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the same structure as \code{dfFileTests}, but
with a new integer column \code{Line}, and one row per line.
}
\description{
Stretch LineStart and LineEnd to Line rows
}
\keyword{internal}
