% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{ExpandTestResultsByIssue}
\alias{ExpandTestResultsByIssue}
\title{Unnest test results}
\usage{
ExpandTestResultsByIssue(dfTestResults)
}
\arguments{
\item{dfTestResults}{(\code{qcthat_TestResults} or data frame) Data frame of test
results as returned by \code{\link[=CompileTestResults]{CompileTestResults()}}.}
}
\value{
A tibble with \code{"Issues"} from \code{\link[=AsTestResultsDF]{AsTestResultsDF()}} unnested into
\code{"Issue"}, and the \code{"Issue"} column first.
}
\description{
Unnest test results
}
\keyword{internal}
