% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTestResults.R
\name{ExtractDisposition}
\alias{ExtractDisposition}
\title{Extract disposition information from a single testthat result}
\usage{
ExtractDisposition(lTestResult)
}
\arguments{
\item{lTestResult}{A single element from a \code{testthat_results} object.}
}
\value{
The string \code{"fail"}, \code{"warn"}, \code{"skip"}, or \code{"pass"}.
}
\description{
Extract disposition information from a single testthat result
}
\keyword{internal}
