% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{ExtractGHClosingIssues}
\alias{ExtractGHClosingIssues}
\title{Extract closing issue numbers from commit messages}
\usage{
ExtractGHClosingIssues(
  chrCommitMessages,
  chrKeywords = c("close", "closes", "closed", "fix", "fixes", "fixed", "resolve",
    "resolves", "resolved"),
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]]
)
}
\arguments{
\item{chrKeywords}{(\code{character}) Keywords to search for just before issue
numbers in commit messages. Defaults to the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub issue-linking keywords}}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}
}
\value{
An integer vector of issue numbers found in \code{chrCommitMessages} that
will be closed according to GitHub's keywords for linking issues to pull
requests.
}
\description{
Extract closing issue numbers from commit messages
}
\keyword{internal}
