% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{ExtractLabels}
\alias{ExtractLabels}
\title{Extract label names from label objects}
\usage{
ExtractLabels(lLabels)
}
\arguments{
\item{lLabels}{(\code{list}) List column of label objects as returned by GitHub
API.}
}
\value{
A list column of character vectors of label names.
}
\description{
Extract label names from label objects
}
\keyword{internal}
