% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestFilesToPotentialIssues.R
\name{ExtractLongTestCommits}
\alias{ExtractLongTestCommits}
\title{Extract test-commit pairs in long format}
\usage{
ExtractLongTestCommits(strTestDir = "tests/testthat", dfFileTests = NULL)
}
\arguments{
\item{strTestDir}{(\verb{length-1 character}) Path to the directory containing
test files. Defaults to \code{"tests/testthat"}.}

\item{dfFileTests}{(\code{data.frame}) A \code{\link[tibble:tibble]{tibble::tibble()}} with the structure
returned by \code{\link[=ExtractTestsFromFiles]{ExtractTestsFromFiles()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per test-commit pair, containing
columns \code{Test}, \code{File}, \code{LineStart}, \code{LineEnd}, \code{Issues}, and \code{Commits}.
}
\description{
Extract test-commit pairs in long format
}
\keyword{internal}
