% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{ExtractName}
\alias{ExtractName}
\title{Extract label names from label objects}
\usage{
ExtractName(lColumn, strName)
}
\arguments{
\item{lColumn}{(\code{list}) A list column of objects.}

\item{strName}{(\verb{length-1 character}) Name of the field to extract from each
object.}
}
\value{
A character vector of the target field from each object, or \code{NA} if
the field is missing from a given element of the list.
}
\description{
Extract label names from label objects
}
\keyword{internal}
