% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{ExtractQcthatCommentID}
\alias{ExtractQcthatCommentID}
\title{Extract the qcthat-comment-id from a comment body}
\usage{
ExtractQcthatCommentID(strBody)
}
\arguments{
\item{strBody}{(\verb{length-1 character}) The body of an issue, PR, or comment,
in GitHub markdown.}
}
\value{
A character vector with the extracted \code{qcthat-comment-id}, or \code{NA}.
}
\description{
Extract the qcthat-comment-id from a comment body
}
\keyword{internal}
