% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTestsFromFiles.R
\name{ExtractTestIssues}
\alias{ExtractTestIssues}
\title{Extract issue numbers from test descriptions}
\usage{
ExtractTestIssues(chrTests)
}
\arguments{
\item{chrTests}{(\code{character}) A vector of test descriptions from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix or extracted from test files.}
}
\value{
A list of integer vectors of sorted unique issue numbers.
}
\description{
Extract issue numbers from test descriptions
}
\keyword{internal}
