% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTestsFromFiles.R
\name{ExtractTestsFromFile}
\alias{ExtractTestsFromFile}
\title{Extract test information from a single file}
\usage{
ExtractTestsFromFile(strFilePath)
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the same structure as
\code{\link[=ExtractTestsFromFiles]{ExtractTestsFromFiles()}}.
}
\description{
Extract test information from a single file
}
\keyword{internal}
