% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{ExtractTestWithPreamble}
\alias{ExtractTestWithPreamble}
\title{Extract test code with file preamble}
\usage{
ExtractTestWithPreamble(
  chrFileLines,
  intLineStart,
  intLineEnd,
  intFirstTestLine
)
}
\arguments{
\item{chrFileLines}{(\code{character}) All lines of the test file.}

\item{intLineStart}{(\verb{length-1 integer}) Start line of the test block.}

\item{intLineEnd}{(\verb{length-1 integer}) End line of the test block.}

\item{intFirstTestLine}{(\verb{length-1 integer}) Line number of the first
\code{test_that()} call in the file.}
}
\value{
A character vector of the preamble (if any) followed by the test
code.
}
\description{
Extract test code with file preamble
}
\keyword{internal}
