% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllIssuePRRefs.R
\name{FetchAllIssuePRRefs}
\alias{FetchAllIssuePRRefs}
\title{Fetch all PRs that reference a specific issue}
\usage{
FetchAllIssuePRRefs(
  intIssue,
  strPRState = c("open", "closed", "merged"),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strPRState}{(\code{character}) States to include in the fetch. Valid values
are \code{"open"}, \code{"closed"}, and \code{"merged"}. By default all states are allowed.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A tibble containing PR details (PR number, State, HeadRef, SHA).
}
\description{
Fetch all PRs that reference a specific issue
}
\keyword{internal}
