% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllIssuePRRefs.R
\name{FetchAllIssuePRRefsRaw}
\alias{FetchAllIssuePRRefsRaw}
\title{Fetch the raw GraphQL response for PRs referencing an issue}
\usage{
FetchAllIssuePRRefsRaw(
  intIssue,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list containing the raw GraphQL response.
}
\description{
Fetch the raw GraphQL response for PRs referencing an issue
}
\keyword{internal}
