% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{FetchAllMergePRNumbers}
\alias{FetchAllMergePRNumbers}
\title{Fetch all PR numbers associated with a vector of commit SHAs}
\usage{
FetchAllMergePRNumbers(
  chrCommitSHAs,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{chrCommitSHAs}{(\code{character}) SHAs of git commits.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A sorted, unique integer vector of associated PR numbers.
}
\description{
Fetch all PR numbers associated with a vector of commit SHAs
}
\keyword{internal}
