% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{FetchAllMergePRNumbersRaw}
\alias{FetchAllMergePRNumbersRaw}
\title{Fetch associated PR data for commits via GraphQL}
\usage{
FetchAllMergePRNumbersRaw(
  chrCommitSHAs,
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{chrCommitSHAs}{(\code{character}) SHAs of git commits.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A raw list response from the \code{\link[gh:gh_gql]{gh::gh_gql()}} call.
}
\description{
Fetch associated PR data for commits via GraphQL
}
\keyword{internal}
