% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllPRCommitSHAs.R
\name{FetchAllPRCommitSHAs}
\alias{FetchAllPRCommitSHAs}
\title{Fetch commit SHAs for multiple PRs using GitHub GraphQL API}
\usage{
FetchAllPRCommitSHAs(
  intPRNumbers,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumbers}{(\code{integer}) A vector of PR numbers.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list of character vectors, one per element of \code{intPRNumbers}, each
containing the commit SHAs for that PR.
}
\description{
Fetches all commits for each PR using batched GraphQL queries so that the
number of API calls scales with the number of PRs divided by the batch size,
not the number of PRs.
}
\keyword{internal}
