% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllPRIssueNumbers.R
\name{FetchAllPRIssueNumbers}
\alias{FetchAllPRIssueNumbers}
\title{Fetch all issue numbers associated with a vector of PRs}
\usage{
FetchAllPRIssueNumbers(
  intPRNumbers,
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumbers}{(\code{integer}) A vector of pull request numbers.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A sorted, unique integer vector of associated issue numbers.
}
\description{
Fetch all issue numbers associated with a vector of PRs
}
\keyword{internal}
