% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{FetchAllPRIssueNumbersRaw}
\alias{FetchAllPRIssueNumbersRaw}
\title{Fetch associated issue data for pull requests via GraphQL}
\usage{
FetchAllPRIssueNumbersRaw(
  intPRNumbers,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumbers}{(\code{integer}) A vector of pull request numbers.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A raw list response from the \code{\link[gh:gh_gql]{gh::gh_gql()}} call.
}
\description{
Fetch associated issue data for pull requests via GraphQL
}
\keyword{internal}
