% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGHLabels.R
\name{FetchGHLabels}
\alias{FetchGHLabels}
\title{Fetch GitHub labels as a data frame}
\usage{
FetchGHLabels(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Label}: Label name.
\item \code{Description}: Label description.
\item \code{Color}: Label color as a hex code (e.g., \code{"#444444"}).
}
}
\description{
Fetch GitHub labels as a data frame
}
\keyword{internal}
