% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGHLabels.R
\name{FetchGHLabelsRaw}
\alias{FetchGHLabelsRaw}
\title{Fetch GitHub labels as raw list}
\usage{
FetchGHLabelsRaw(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list of label objects as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch GitHub labels as raw list
}
\keyword{internal}
