% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{FetchGQL}
\alias{FetchGQL}
\title{Wrapper for GitHub GraphQL API calls}
\usage{
FetchGQL(
  strQuery,
  ...,
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strQuery}{(\verb{length-1 character}) The GraphQL query (or sub-query).}

\item{...}{Additional parameters passed to \code{\link[gh:gh_gql]{gh::gh_gql()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The result of the \code{\link[gh:gh_gql]{gh::gh_gql()}} call.
}
\description{
Wrapper for GitHub GraphQL API calls
}
\keyword{internal}
