% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{FetchIssueCommentGHID}
\alias{FetchIssueCommentGHID}
\title{Fetch the GitHub ID of an issue comment by its qcthat comment ID}
\usage{
FetchIssueCommentGHID(
  intIssue,
  strCommentID,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strCommentID}{(\verb{length-1 character}) A unique ID for a comment within a
given context, which is usually a hash of the title of the comment.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The GitHub ID of the comment.
}
\description{
Fetch the GitHub ID of an issue comment by its qcthat comment ID
}
\keyword{internal}
