% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueComments.R
\name{FetchIssueComments}
\alias{FetchIssueComments}
\title{Fetch comments on an issue}
\usage{
FetchIssueComments(
  intIssue,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A \code{qcthat_Comments} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{Body}: Comment body (the full text of the comment).
\item \code{Url}: URL of the comment on GitHub.
\item \code{CommentGHID}: GitHub ID of the comment.
\item \code{Author}: GitHub username of the comment author.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the issue was created.
\item \code{UpdatedAt}: \code{POSIXct} timestamp of when the issue was last updated.
\item \code{qcthatCommentID}: The \code{qcthat-comment-id}, which is usually a hash of
the title of the comment (when present).
}
}
\description{
Fetch comments on an issue
}
\keyword{internal}
