% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueDetails.R
\name{FetchIssueDetails}
\alias{FetchIssueDetails}
\title{Fetch issue details from GitHub}
\usage{
FetchIssueDetails(
  intIssues,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssues}{(\code{integer}) A vector of issue numbers from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix or from GitHub.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{Title}: Issue title.
\item \code{Body}: Issue body (or \code{NA_character_} if NULL).
}
}
\description{
Fetch issue details from GitHub
}
\keyword{internal}
