% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchIssueDetails.R
\name{FetchIssueDetailsRaw}
\alias{FetchIssueDetailsRaw}
\title{Fetch raw issue details from GitHub API}
\usage{
FetchIssueDetailsRaw(
  intIssue,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A \code{list} containing raw issue data from the GitHub API.
}
\description{
Fetch raw issue details from GitHub API
}
\keyword{internal}
