% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{FetchIssueUAChildren}
\alias{FetchIssueUAChildren}
\title{Fetch all user-acceptance sub-issues for an issue}
\usage{
FetchIssueUAChildren(
  intIssue,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A data frame of user-acceptance sub-issues.
}
\description{
Fetch all user-acceptance sub-issues for an issue
}
\keyword{internal}
