% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{FetchJobID}
\alias{FetchJobID}
\title{Fetch the GitHub Actions job ID for a given run ID and job name}
\usage{
FetchJobID(
  strRunID = character(),
  strJobName = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strJobName}{(\verb{length-1 character}) Name of a GitHub Actions workflow
job.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
The GitHub Actions job ID, or \code{NULL} if not found.
}
\description{
Fetch the GitHub Actions job ID for a given run ID and job name
}
\keyword{internal}
