% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{FetchMergeCommitBatchRaw}
\alias{FetchMergeCommitBatchRaw}
\title{Fetch a single page of commits from a comparison}
\usage{
FetchMergeCommitBatchRaw(
  strSourceRef,
  strTargetRef,
  intPage = 1,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{intPage}{(\verb{length-1 integer}) The page number to fetch.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A raw list response from the API.
}
\description{
Fetch a single page of commits from a comparison
}
\keyword{internal}
