% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchMergeIssueNumbers.R
\name{FetchMergeIssueClosers}
\alias{FetchMergeIssueClosers}
\title{Fetch issue closers associated with a merge}
\usage{
FetchMergeIssueClosers(
  intPRNumbers,
  chrCommitSHAs,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumbers}{(\code{integer}) A vector of pull request numbers.}

\item{chrCommitSHAs}{(\code{character}) SHAs of git commits.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A sorted, unique integer vector of associated issue numbers.
}
\description{
Fetch issue closers associated with a merge
}
\keyword{internal}
