% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchMergeIssueNumbers.R
\name{FetchMergeIssueNumbers}
\alias{FetchMergeIssueNumbers}
\title{Fetch all issue numbers associated with a merge}
\usage{
FetchMergeIssueNumbers(
  strSourceRef = GetActiveBranch(),
  strTargetRef = GetDefaultBranch(),
  intPageMax = 100L,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{intPageMax}{(\verb{length-1 integer}) The maximum number of pages of commits
to fetch from the GitHub API. Each page contains up to 100 commits.
Defaults to 100, which fetches up to 10,000 commits. You likely never need
to increase this number, but try a larger number if a merge involves a very
large number of commits in a very large repository.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A sorted, unique integer vector of associated issue numbers.
}
\description{
Fetch all issue numbers associated with a merge
}
\keyword{internal}
