% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchAllPRCommitSHAs.R
\name{FetchPRCommitSHAsBatch}
\alias{FetchPRCommitSHAsBatch}
\title{Fetch commit SHAs for a batch of PRs in one GraphQL query}
\usage{
FetchPRCommitSHAsBatch(
  intPRNumbers,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumbers}{(\code{integer}) A batch of PR numbers to query together.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list of character vectors, one per element of \code{intPRNumbers}.
}
\description{
Fetch commit SHAs for a batch of PRs in one GraphQL query
}
\keyword{internal}
