% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{FetchPRUATCommentIssues}
\alias{FetchPRUATCommentIssues}
\title{Extract issue numbers from a PR's UAT comment}
\usage{
FetchPRUATCommentIssues(
  intPRNumber,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A sorted integer vector of unique issue numbers found in the UAT
comment. Returns \code{NULL} invisibly if no comment or no issues are found.
}
\description{
Extract issue numbers from a PR's UAT comment
}
\keyword{internal}
