% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessReleaseID.R
\name{FetchRawReleaseByTagName}
\alias{FetchRawReleaseByTagName}
\title{Fetch a release from GitHub}
\usage{
FetchRawReleaseByTagName(
  strTagName,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strTagName}{(\verb{length-1 character}) Name of the GitHub tag.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A raw pull request object as a list as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch a release from GitHub
}
\keyword{internal}
