% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{FetchRawRepoIssues}
\alias{FetchRawRepoIssues}
\title{Fetch all repo issues from GitHub}
\usage{
FetchRawRepoIssues(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token(),
  strState = c("all", "open", "closed")
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{strState}{(\verb{length-1 character}) State of issues or pull requests to
fetch. Must be one of \code{"open"}, \code{"closed"}, or \code{"all"}. Defaults to
\code{"open"} for pull requests and \code{"all"} for issues.}
}
\value{
A list of raw issue objects as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch all repo issues from GitHub
}
\keyword{internal}
