% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{FetchRawRepoPRs}
\alias{FetchRawRepoPRs}
\title{Fetch repo PRs from GitHub}
\usage{
FetchRawRepoPRs(
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token(),
  strState = c("open", "closed", "all")
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{strState}{(\verb{length-1 character}) State of issues or pull requests to
fetch. Must be one of \code{"open"}, \code{"closed"}, or \code{"all"}. Defaults to
\code{"open"} for pull requests and \code{"all"} for issues.}
}
\value{
A list of raw pull request objects as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch repo PRs from GitHub
}
\keyword{internal}
