% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchPRIssueNumbers.R
\name{FetchRawRepoPRSingle}
\alias{FetchRawRepoPRSingle}
\title{Fetch a single repo PR from GitHub}
\usage{
FetchRawRepoPRSingle(
  intPRNumber = GuessPRNumber(".", strOwner, strRepo, strGHToken),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A list representing a raw pull request object as returned by
\code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch a single repo PR from GitHub
}
\keyword{internal}
