% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{FetchRefPRNumber}
\alias{FetchRefPRNumber}
\title{Fetch the pull request number for a branch or other git ref}
\usage{
FetchRefPRNumber(
  strSourceRef = GetActiveBranch(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
An integer pull request number, or \code{integer()} if no matching PR is
found. If multiple PRs are found, first, if there are any open PRs, we
filter to only include open PRs, then the PR that was most recently created
is returned.
}
\description{
Fetch the pull request number for a branch or other git ref
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  FetchRefPRNumber()
\dontshow{\}) # examplesIf}
}
