% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssueClosers.R
\name{FetchRepoIssueClosers}
\alias{FetchRepoIssueClosers}
\title{Fetch repository issue closers}
\usage{
FetchRepoIssueClosers(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{CloserType}: Type of the closer, either \code{Commit} or \code{PullRequest}.
\item \code{CloserSHA}: SHA of the commit that closed the issue. For \code{Commit}
closers, this is the commit OID directly. For \code{PullRequest} closers, this
is the merge commit SHA.
\item \code{CloserPRNumber}: Number of the pull request that closed the issue, or
\code{NA} if the issue was closed by a commit.
}
}
\description{
Fetch the closers (commits or pull requests) for all closed issues in a
repository.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  FetchRepoIssueClosers()
\dontshow{\}) # examplesIf}
}
