% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssueClosers.R
\name{FetchRepoIssueClosersRaw}
\alias{FetchRepoIssueClosersRaw}
\title{Fetch raw repository issue closers}
\usage{
FetchRepoIssueClosersRaw(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list of raw issue closer objects as returned by \code{\link[gh:gh_gql]{gh::gh_gql()}}.
}
\description{
Fetch raw repository issue closers
}
\keyword{internal}
