% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{FetchRepoIssues}
\alias{FetchRepoIssues}
\title{Fetch repository issues}
\usage{
FetchRepoIssues(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token(),
  strState = c("all", "open", "closed")
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{strState}{(\verb{length-1 character}) State of issues or pull requests to
fetch. Must be one of \code{"open"}, \code{"closed"}, or \code{"all"}. Defaults to
\code{"open"} for pull requests and \code{"all"} for issues.}
}
\value{
A \code{qcthat_Issues} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{Title}: Issue title.
\item \code{Body}: Issue body (the full text of the issue).
\item \code{Labels}: List column of character vectors of issue labels.
\item \code{State}: Issue state (\code{open} or \code{closed}).
\item \code{StateReason}: Reason for issue state (e.g., \code{completed}) or \code{NA} if not
applicable.
\item \code{Milestone}: Issue milestone title or \code{NA} if not applicable.
\item \code{Type}: Issue type or \code{"Issue"} if no issue type is available.
\item \code{Url}: URL of the issue on GitHub.
\item \code{ParentOwner}: GitHub username or organization name of the parent issue
if applicable, otherwise \code{NA}.
\item \code{ParentRepo}: GitHub repository name of the parent issue if applicable,
otherwise \code{NA}.
\item \code{ParentNumber}: GitHub issue number of the parent issue if applicable,
otherwise \code{NA}.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the issue was created.
\item \code{ClosedAt}: \code{POSIXct} timestamp of when the issue was closed, or \code{NA} if
the issue is still open.
}
}
\description{
Download (non-pull-request) issues in a repository and parse them into a tidy
\code{\link[tibble:tibble]{tibble::tibble()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  FetchRepoIssues()
\dontshow{\}) # examplesIf}
}
