% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{FetchRepoPRs}
\alias{FetchRepoPRs}
\title{Fetch repository pull requests}
\usage{
FetchRepoPRs(
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token(),
  strState = c("open", "closed", "all")
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{strState}{(\verb{length-1 character}) State of issues or pull requests to
fetch. Must be one of \code{"open"}, \code{"closed"}, or \code{"all"}. Defaults to
\code{"open"} for pull requests and \code{"all"} for issues.}
}
\value{
A \code{qcthat_PRs} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{PR}: Pull request number.
\item \code{Title}: Pull request title.
\item \code{State}: Pull request state (\code{"open"} or \code{"closed"}).
\item \code{HeadRef}: The head reference (branch) of the pull request (the changes).
\item \code{BaseRef}: The base reference (branch) of the pull request (the target).
\item \code{Body}: The full text of the pull request.
\item \code{MergeCommitSHA}: The SHA of the merge commit, if the pull request has been
merged.
\item \code{IsDraft}: Logical indicating whether the pull request is a draft.
\item \code{Url}: URL of the pull request on GitHub.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the pull request was created.
\item \code{ClosedAt}: \code{POSIXct} timestamp of when the pull request was closed, or
\code{NA} if the pull request is still open.
\item \code{MergedAt}: \code{POSIXct} timestamp of when the pull request was merged, or
\code{NA} if the pull request has not been merged.
}
}
\description{
Download pull requests in a repository and parse them into a tidy
\code{\link[tibble:tibble]{tibble::tibble()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  FetchRepoPRs()
\dontshow{\}) # examplesIf}
}
