% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{FetchRunInfoRaw}
\alias{FetchRunInfoRaw}
\title{Fetch raw run info for a GitHub Actions run}
\usage{
FetchRunInfoRaw(
  strRunID,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A list with the raw run object as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch raw run info for a GitHub Actions run
}
\keyword{internal}
