% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentReport.R
\name{FetchRunJobsRaw}
\alias{FetchRunJobsRaw}
\title{Fetch raw job info for a GitHub Actions run}
\usage{
FetchRunJobsRaw(
  strRunID,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A list of raw job objects as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch raw job info for a GitHub Actions run
}
\keyword{internal}
