% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{FetchRunURL}
\alias{FetchRunURL}
\title{Fetch the URL for a GitHub Actions run}
\usage{
FetchRunURL(
  strRunID = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A string containing the URL for the specified GitHub Actions run,
linking to the specific job when there is only one job.
}
\description{
Fetch the URL for a GitHub Actions run
}
\keyword{internal}
