% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{FetchSessionInfoStepNumber}
\alias{FetchSessionInfoStepNumber}
\title{Fetch the session info step number for a GitHub Actions job}
\usage{
FetchSessionInfoStepNumber(
  strJobID,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strJobID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run job.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
The step number of the session info step, or \code{NULL} if not found.
}
\description{
Fetch the session info step number for a GitHub Actions job
}
\keyword{internal}
