% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchUAIssue.R
\name{FetchUAIssue}
\alias{FetchUAIssue}
\title{Fetch or create a user-acceptance issue}
\usage{
FetchUAIssue(
  strDescription,
  intIssue,
  chrChecks = character(),
  chrInstructions = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strDescription}{(\verb{length-1 character}) A brief description of a user
expectation.}

\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{chrChecks}{(\code{character}) Items for the user to check. These will be
preceded by checkboxes in the associated issue.}

\item{chrInstructions}{(\code{character}) Instructions for how to review an issue.
Included in the associated issue before the checklist.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A list representing the user-acceptance issue.
}
\description{
Fetch or create a user-acceptance issue
}
\keyword{internal}
