% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{FetchVectorFromGQL}
\alias{FetchVectorFromGQL}
\title{Fetch a vector of values from GitHub GraphQL API using a builder function}
\usage{
FetchVectorFromGQL(
  x,
  fnBuildQuery,
  vecProto = integer(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{x}{(\code{vector}) A vector of inputs to process.}

\item{fnBuildQuery}{(\code{function}) A function that takes a single element of
\code{x} and returns a character string representing a GraphQL sub-query for
that element.}

\item{vecProto}{(\code{vector}) A prototype vector to return if \code{x} is empty.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A sorted, unique vector of values fetched from GitHub, or
\code{vecProto}.
}
\description{
Fetch a vector of values from GitHub GraphQL API using a builder function
}
\keyword{internal}
