% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{FileExists}
\alias{FileExists}
\title{Wrapper for test mocking}
\usage{
FileExists(strPath)
}
\arguments{
\item{strPath}{(\verb{length-1 character}) Path to check for existence.}
}
\value{
(\code{logical}) Whether the file exists.
}
\description{
Wrapper for test mocking
}
\keyword{internal}
