% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{FinalizeTree}
\alias{FinalizeTree}
\title{Finalize a qcthat tree}
\usage{
FinalizeTree(lTree)
}
\arguments{
\item{lTree}{(\code{list} of \code{character}) A list of character vectors representing
parts of the qcthat report.}
}
\value{
A character vector representing the finalized tree structure.
}
\description{
Finalize a qcthat tree
}
\keyword{internal}
