% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRepoIssuesToCommits.R
\name{FindAllIssueCommits}
\alias{FindAllIssueCommits}
\title{Add Commits list column to dfIssueClosers}
\usage{
FindAllIssueCommits(
  strCloserType,
  strCloserSHA,
  intCloserPRNumber,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token(),
  lPRs = NULL
)
}
\arguments{
\item{strCloserType}{(\verb{length-1 character}) Whether the issue was closed by a
\code{"PullRequest"} or a \code{"Commit"}.}

\item{strCloserSHA}{(\verb{length-1 character}) The commit SHA for an issue closed
by a \code{"Commit"}.}

\item{intCloserPRNumber}{(\verb{length-1 integer}) The number of the pull request
that closed the issue.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{lPRs}{(\code{list} or \code{NULL}) Optional list of raw pull request objects as
returned by \code{\link[=FetchRawRepoPRs]{FetchRawRepoPRs()}}. If provided, PRs will be looked up from
this list instead of fetching individually from the API.}
}
\value{
A character vector of commit SHAs associated with this issue.
}
\description{
Add Commits list column to dfIssueClosers
}
