% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{FindFirstTestLine}
\alias{FindFirstTestLine}
\title{Find the line number of the first test_that call}
\usage{
FindFirstTestLine(chrFileLines)
}
\arguments{
\item{chrFileLines}{(\code{character}) Lines of a test file.}
}
\value{
The line number of the first \verb{test_that(} call, or
\code{length(chrFileLines) + 1L} if no tests are found.
}
\description{
Find the line number of the first test_that call
}
\keyword{internal}
