% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{FindKeywordIssues}
\alias{FindKeywordIssues}
\title{Find issues that will be closed by merging one branch into another}
\usage{
FindKeywordIssues(
  strSourceRef = GetActiveBranch(strPkgRoot),
  strTargetRef = GetDefaultBranch(strPkgRoot),
  strPkgRoot = ".",
  chrKeywords = c("close", "closes", "closed", "fix", "fixes", "fixed", "resolve",
    "resolves", "resolved"),
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  intMaxCommits = 100000L
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{chrKeywords}{(\code{character}) Keywords to search for just before issue
numbers in commit messages. Defaults to the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub issue-linking keywords}}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{intMaxCommits}{(\verb{length-1 integer}) The maximum number of commits to
return from git logs. Leaving this at the default should almost always be
fine, but you can reduce the number if your repository has a long commit
history and this function is slow.}
}
\value{
An integer vector of issue numbers in this repo that will be closed
by merging \code{strSourceRef} into \code{strTargetRef}, using the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub keywords for linking issues to pull requests}.
}
\description{
Find issues that will be closed by merging one branch into another
}
\keyword{internal}
