% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTestsFromFiles.R
\name{FindTests}
\alias{FindTests}
\title{Find test_that calls in a file}
\usage{
FindTests(chrTestLines)
}
\arguments{
\item{chrTestLines}{(\code{character}) Vector of lines from a test file.}
}
\value{
A list of lists, each containing \code{desc}, \code{start}, and \code{end}.
}
\description{
Find test_that calls in a file
}
\keyword{internal}
