% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R, R/PrintIssueTestMatrix.R,
%   R/PrintMilestone.R, R/PrintSingleIssueTestResults.R
\name{FormatBody}
\alias{FormatBody}
\alias{FormatBody.default}
\alias{FormatBody.qcthat_IssueTestMatrix}
\alias{FormatBody.qcthat_Milestone}
\alias{FormatBody.qcthat_SingleIssueTestResults}
\title{Format the body of a qcthat object}
\usage{
FormatBody(x, ..., lglUseEmoji = getOption("qcthat.emoji", TRUE))

\method{FormatBody}{default}(x, ..., lglUseEmoji = getOption("qcthat.emoji", TRUE))

\method{FormatBody}{qcthat_IssueTestMatrix}(x, ..., lglShowMilestones = TRUE)

\method{FormatBody}{qcthat_Milestone}(x, ...)

\method{FormatBody}{qcthat_SingleIssueTestResults}(x, ..., lglUseEmoji = getOption("qcthat.emoji", TRUE))
}
\arguments{
\item{x}{(\code{qcthat_Object}) The qcthat object to format.}

\item{...}{Additional arguments passed to methods.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}

\item{lglShowMilestones}{(\verb{length-1 logical}) Whether to separate issues by
milestones in reports.}
}
\value{
A character vector representing the formatted body of the object. By
default, if no method is defined for the object, the body is \code{\link[=format]{format()}}
applied to the object with the \code{"qcthat_object"} class removed.
}
\description{
Format the body of a qcthat object
}
\keyword{internal}
