% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{FormatFooterMain}
\alias{FormatFooterMain}
\title{Format the non-key portion of the footer}
\usage{
FormatFooterMain(
  x,
  ...,
  lglUseEmoji = getOption("qcthat.emoji", TRUE),
  lglShowIgnoredLabels = TRUE
)
}
\arguments{
\item{x}{(\code{qcthat_Object}) The qcthat object to format.}

\item{...}{Additional arguments passed to methods.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}

\item{lglShowIgnoredLabels}{(\verb{length-1 logical}) Whether to show information
in reports about issue labels (such as \code{"qcthat-nocov"}) that have been
ignored.}
}
\value{
A character vector representing the formatted footer main body.
}
\description{
This is a placeholder. Ideally it should probably also be an S3 generic, or
otherwise have a more specific name.
}
\keyword{internal}
