% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R, R/PrintIssueTestMatrix.R,
%   R/PrintMilestone.R, R/PrintSingleIssueTestResults.R
\name{FormatHeader}
\alias{FormatHeader}
\alias{FormatHeader.default}
\alias{FormatHeader.qcthat_IssueTestMatrix}
\alias{FormatHeader.qcthat_Milestone}
\alias{FormatHeader.qcthat_SingleIssueTestResults}
\title{Format the header of a qcthat object}
\usage{
FormatHeader(x, ..., lglUseEmoji = getOption("qcthat.emoji", TRUE))

\method{FormatHeader}{default}(x, ...)

\method{FormatHeader}{qcthat_IssueTestMatrix}(x, ..., lglUseEmoji = getOption("qcthat.emoji", TRUE))

\method{FormatHeader}{qcthat_Milestone}(x, ...)

\method{FormatHeader}{qcthat_SingleIssueTestResults}(x, ..., lglUseEmoji = getOption("qcthat.emoji", TRUE))
}
\arguments{
\item{x}{(\code{qcthat_Object}) The qcthat object to format.}

\item{...}{Additional arguments passed to methods.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A character vector representing the formatted header of the object.
By default, if no method is defined for the object, no header is printed.
}
\description{
Format the header of a qcthat object
}
\keyword{internal}
