% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatIssue}
\alias{FormatIssue}
\title{Prepare a single issue of a milestone for the qcthat_IssueTestMatrix output}
\usage{
FormatIssue(Issue, Title, StateIndicator, Type, TestResults, ...)
}
\arguments{
\item{Issue}{(\verb{length-1 integer}) The issue number.}

\item{Title}{(\verb{length-1 character}) The issue title.}

\item{StateIndicator}{(\verb{length-1 character}) The issue state indicator.}

\item{Type}{(\verb{length-1 character}) The issue type.}

\item{TestResults}{(\code{data.frame}) A data frame with test results for the
issue.}

\item{...}{Additional arguments (not used).}
}
\value{
A character vector representing the issue and its tests.
}
\description{
Prepare a single issue of a milestone for the qcthat_IssueTestMatrix output
}
\keyword{internal}
