% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatIssueHeader}
\alias{FormatIssueHeader}
\title{Prepare the issue itself for the qcthat_IssueTestMatrix output}
\usage{
FormatIssueHeader(Issue, StateIndicator, Title, Type)
}
\arguments{
\item{Issue}{(\verb{length-1 integer}) The issue number.}

\item{StateIndicator}{(\verb{length-1 character}) The issue state indicator.}

\item{Title}{(\verb{length-1 character}) The issue title.}

\item{Type}{(\verb{length-1 character}) The issue type.}
}
\value{
A length-1 character vector representing the issue header.
}
\description{
Prepare the issue itself for the qcthat_IssueTestMatrix output
}
\keyword{internal}
