% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatIssues}
\alias{FormatIssues}
\title{Prepare the issues of a milestone for the qcthat_IssueTestMatrix output}
\usage{
FormatIssues(IssueTestResults)
}
\arguments{
\item{IssueTestResults}{(\code{qcthat_IssueTestResults}) A data frame with issue
test results for the milestone.}
}
\value{
A character vector representing the issues and their tests.
}
\description{
Prepare the issues of a milestone for the qcthat_IssueTestMatrix output
}
\keyword{internal}
