% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{FormatKey}
\alias{FormatKey}
\title{Format the key for interpreting issue and test states}
\usage{
FormatKey(lglUseEmoji = getOption("qcthat.emoji", TRUE))
}
\arguments{
\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A character vectore styled with \code{\link[pillar:style_subtle]{pillar::style_subtle()}}.
}
\description{
This is a placeholder. Eventually we should use dfITM to determine which
pieces of the key should be included.
}
\keyword{internal}
