% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatMilestone}
\alias{FormatMilestone}
\title{Prepare a single milestone of the qcthat_IssueTestMatrix output}
\usage{
FormatMilestone(Milestone, IssueTestResults)
}
\arguments{
\item{Milestone}{(\verb{length-1 character}) The name of the milestone.}

\item{IssueTestResults}{(\code{qcthat_IssueTestResults}) A data frame with issue
test results for the milestone.}
}
\value{
If \code{x} has rows, a character vector representing the tree that
describes the milestones, issues, and tests in \code{x}. \code{NULL} otherwise.
}
\description{
Prepare a single milestone of the qcthat_IssueTestMatrix output
}
\keyword{internal}
