% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatMilestones}
\alias{FormatMilestones}
\title{Prepare the main part of the qcthat_IssueTestMatrix output}
\usage{
FormatMilestones(x)
}
\arguments{
\item{x}{(\code{qcthat_IssueTestMatrix}) The issue-test matrix to format.}
}
\value{
If \code{x} has rows, a character vector representing the tree that
describes the milestones, issues, and tests in \code{x}. \code{NULL} otherwise.
}
\description{
Prepare the main part of the qcthat_IssueTestMatrix output
}
\keyword{internal}
