% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentAllReports.R
\name{FormatReportBody}
\alias{FormatReportBody}
\title{Format the body of a report}
\usage{
FormatReportBody(strReportType, strBody)
}
\arguments{
\item{strReportType}{(\verb{length-1 character}) The main title of the report,
such as \code{"Completed Issues"} or \code{"PR-Associated Issues"}.}

\item{strBody}{(\verb{length-1 character}) The body of the report.}
}
\value{
A string containing the report formatted in GitHub markdown.
}
\description{
Format the body of a report
}
\keyword{internal}
