% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentReport.R
\name{FormatReportGH}
\alias{FormatReportGH}
\title{Format a QC report in GitHub markdown}
\usage{
FormatReportGH(dfITM)
}
\arguments{
\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}}.}
}
\value{
A string containing the report formatted in GitHub markdown.
}
\description{
Format a QC report in GitHub markdown
}
\keyword{internal}
