% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentAllReports.R
\name{FormatReportType}
\alias{FormatReportType}
\title{Prepare the body of one report in set of reports}
\usage{
FormatReportType(
  fnReport,
  strReportType,
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  chrIgnoredLabels = DefaultIgnoreLabels(),
  dfITM = NULL,
  lOtherArgs = list(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{fnReport}{(\code{function}) The function that generates the report data,
such as \code{\link[=QCPR]{QCPR()}}.}

\item{strReportType}{(\verb{length-1 character}) The main title of the report,
such as \code{"Completed Issues"} or \code{"PR-Associated Issues"}.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}} (often via \code{\link[=QCPackage]{QCPackage()}}).}

\item{lOtherArgs}{(\code{list}) Additional arguments to pass to \code{fnReport}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A string containing the report formatted in GitHub markdown.
}
\description{
Prepare the body of one report in set of reports
}
\keyword{internal}
