% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentReport.R
\name{FormatSessionInfo}
\alias{FormatSessionInfo}
\title{Format session info in GitHub markdown}
\usage{
FormatSessionInfo()
}
\value{
A string containing the session info formatted in GitHub markdown.
}
\description{
Format session info in GitHub markdown
}
\keyword{internal}
