% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSingleIssueTestResults.R
\name{FormatSITRHeader}
\alias{FormatSITRHeader}
\title{Format a non-empty SingleIssueTestResults header}
\usage{
FormatSITRHeader(x, lglUseEmoji = getOption("qcthat.emoji", TRUE))
}
\arguments{
\item{x}{(\code{qcthat_Object}) The qcthat object to format.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A formatted string for the SingleIssueTestResults header.
}
\description{
Format a non-empty SingleIssueTestResults header
}
\keyword{internal}
