% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatTestResults}
\alias{FormatTestResults}
\title{Prepare the test results for the qcthat_IssueTestMatrix output}
\usage{
FormatTestResults(TestResults)
}
\arguments{
\item{TestResults}{(\code{data.frame}) A data frame with test results for the
issue.}
}
\value{
A character vector representing the issue's test results.
}
\description{
Prepare the test results for the qcthat_IssueTestMatrix output
}
\keyword{internal}
