% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{formatting}
\alias{formatting}
\alias{print.qcthat_IssueTestMatrix}
\alias{format.qcthat_IssueTestMatrix}
\title{Printing Issue-Test Matrices}
\usage{
\method{print}{qcthat_IssueTestMatrix}(x, ...)

\method{format}{qcthat_IssueTestMatrix}(x, ..., transform = identity)
}
\arguments{
\item{x}{(\code{qcthat_IssueTestMatrix}) The issue-test matrix to format.}

\item{...}{These dots are for future extensions and must be empty.}

\item{transform}{(\code{function}) A function to transform the output before
returning it. By default, this is \code{\link[base:identity]{base::identity()}} for \code{format()}
(returns the formatted output), and \code{\link[base:writeLines]{base::writeLines()}} for \code{print()}
(sends the output to \code{\link[=stdout]{stdout()}}).}
}
\description{
One of the main purposes of qcthat is to produce user-friendly reports about
a package's issue and test status. Printing an issue-test matrix provides a
nested view of the relationship between milestones, issues, and tests, as
well as the state of issues ("open", "closed (completed)", or "closed (won't
fix)") and the disposition of tests ("passed", "failed", or "skipped").
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  lTestResults <- testthat::test_local(stop_on_failure = FALSE)
  IssueTestMatrix <- CompileIssueTestMatrix( dfRepoIssues =
  FetchRepoIssues(), dfTestResults = CompileTestResults(lTestResults) )
  print(IssueTestMatrix)

  # Remove the "qcthat_IssueTestMatrix" class to see the raw tibble.
  print(tibble::as_tibble(IssueTestMatrix))
\dontshow{\}) # examplesIf}
}
