% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentUAT.R
\name{FormatUATGH}
\alias{FormatUATGH}
\title{Format UAT issues in GitHub markdown}
\usage{
FormatUATGH(intIssues = integer())
}
\arguments{
\item{intIssues}{(\code{integer}) A vector of issue numbers from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix or from GitHub.}
}
\value{
A string containing the UAT issues formatted in GitHub markdown.
}
\description{
Format UAT issues in GitHub markdown
}
\keyword{internal}
