% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{GetActiveBranch}
\alias{GetActiveBranch}
\title{Determine the active branch of a git repository}
\usage{
GetActiveBranch(strPkgRoot = ".")
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
A length-1 character vector representing the active branch name.
}
\description{
Determine the active branch of a git repository
}
\keyword{internal}
