% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{GetDefaultBranch}
\alias{GetDefaultBranch}
\title{Determine the default branch of a git repository}
\usage{
GetDefaultBranch(strPkgRoot = ".")
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
A length-1 character vector representing the default branch name.
}
\description{
Determine the default branch of a git repository
}
\keyword{internal}
