% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{GetGHAPRNumber}
\alias{GetGHAPRNumber}
\title{Get the PR number for a GitHub Action}
\usage{
GetGHAPRNumber(lGHEventPayload = LoadGHEventPayload())
}
\arguments{
\item{lGHEventPayload}{(\code{list}) The GitHub event payload. Defaults to the
result of \code{\link[=LoadGHEventPayload]{LoadGHEventPayload()}}.}
}
\value{
An integer pull request number, or \code{NULL} if the GitHub Actions
event payload does not include a pull request number (for example, when
the workflow was not triggered by a pull_request event or no \code{inputs.pr}
value was provided).
}
\description{
Get the PR number for a GitHub Action
}
\keyword{internal}
