% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{GetGHRemote}
\alias{GetGHRemote}
\title{Find the target repository}
\usage{
GetGHRemote(strPkgRoot = ".")
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
A list representing the GitHub repository at \code{strPkgRoot}.
}
\description{
A wrapper to safely call \code{\link[gh:gh_tree_remote]{gh::gh_tree_remote()}} if the project uses git.
}
\keyword{internal}
