% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{GetGHRemoteList}
\alias{GetGHRemoteList}
\title{Wrapper around gert::git_remote_list() for mocking}
\usage{
GetGHRemoteList(strPkgRoot)
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_find]{gert::git_find()}}.}
}
\value{
The result of the \code{\link[gert:git_remote_list]{gert::git_remote_list()}} call.
}
\description{
Wrapper around gert::git_remote_list() for mocking
}
\keyword{internal}
