% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{GetGHRepo}
\alias{GetGHRepo}
\title{Find the name of the target repository}
\usage{
GetGHRepo(strPkgRoot = ".")
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
A length-1 \code{character} vector representing the GitHub repo name of
the repository at \code{strPkgRoot}.
}
\description{
A wrapper to safely call \code{\link[gh:gh_tree_remote]{gh::gh_tree_remote()}} and extract the name
(\code{"repo"}).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
  GetGHRepo()
\dontshow{\}) # examplesIf}
}
