% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{GetGitAheadBehind}
\alias{GetGitAheadBehind}
\title{Wrapper around gert::git_ahead_behind() for mocking}
\usage{
GetGitAheadBehind(strUpstream, strRef, strPkgRoot = ".")
}
\arguments{
\item{strUpstream}{(\verb{length-1 character}) The upstream branch or commit SHA.}

\item{strRef}{(\verb{length-1 character}) The local branch or commit SHA.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
The result of \code{\link[gert:git_rebase]{gert::git_ahead_behind()}}.
}
\description{
Wrapper around gert::git_ahead_behind() for mocking
}
\keyword{internal}
