% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{GetGitCommitInfo}
\alias{GetGitCommitInfo}
\title{Wrapper around gert::git_commit_info() for mocking}
\usage{
GetGitCommitInfo(strRef, strPkgRoot = ".")
}
\arguments{
\item{strRef}{(\verb{length-1 character}) A branch, tag, or commit SHA.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
The result of \code{\link[gert:git_commit]{gert::git_commit_info()}}.
}
\description{
Wrapper around gert::git_commit_info() for mocking
}
\keyword{internal}
