% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{GetGitLog}
\alias{GetGitLog}
\title{List commits in a git reference}
\usage{
GetGitLog(strGitRef, strPkgRoot = ".", intMaxCommits = 1e+05)
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{intMaxCommits}{(\verb{length-1 integer}) The maximum number of commits to
return from git logs. Leaving this at the default should almost always be
fine, but you can reduce the number if your repository has a long commit
history and this function is slow.}
}
\value{
A data frame of commits that are in \code{strGitRef}, as returned by
\code{\link[gert:git_commit]{gert::git_log()}}.
}
\description{
List commits in a git reference
}
\keyword{internal}
