% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{GetPkgRoot}
\alias{GetPkgRoot}
\title{Find the root of the package repo}
\usage{
GetPkgRoot(strPkgRoot, envCall = rlang::caller_env())
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
The path to the root of the package.
}
\description{
Find the root of the package repo
}
\keyword{internal}
