% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentReport.R
\name{GetRawSessionInfo}
\alias{GetRawSessionInfo}
\title{Get raw session info}
\usage{
GetRawSessionInfo()
}
\value{
The raw session info as returned by either
\code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} (if the \code{sessioninfo} package is installed)
or \code{\link[=sessionInfo]{sessionInfo()}}.
}
\description{
Get raw session info
}
\keyword{internal}
