% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{GetRelativePackagePath}
\alias{GetRelativePackagePath}
\title{Find the path to a file}
\usage{
GetRelativePackagePath(strFilePath, envCall = rlang::caller_env())
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
The path to a file, relative to the root of the package that the
file is within.
}
\description{
Find the path to a file
}
\keyword{internal}
