% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttachReleaseReports.R
\name{GetSessionInfo}
\alias{GetSessionInfo}
\title{Get session info from available reporting functions}
\usage{
GetSessionInfo()
}
\value{
The raw session info as returned by either
\code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} (if the \code{sessioninfo} package is installed)
or \code{\link[=sessionInfo]{sessionInfo()}}.
}
\description{
Retrieve session info using either \code{sessioninfo::session_info()} (if the
\code{sessioninfo} package is installed) or \code{utils::sessionInfo()}. The former
provides more detailed information about the R session and installed
packages.
}
\examples{
GetSessionInfo()
}
