% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindKeywordIssues.R
\name{GHKeywordRegex}
\alias{GHKeywordRegex}
\title{Compile the regex used to find GitHub-closing issues in commit messages}
\usage{
GHKeywordRegex(
  chrKeywords = c("close", "closes", "closed", "fix", "fixes", "fixed", "resolve",
    "resolves", "resolved"),
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]]
)
}
\arguments{
\item{chrKeywords}{(\code{character}) Keywords to search for just before issue
numbers in commit messages. Defaults to the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub issue-linking keywords}}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}
}
\value{
A length-1 character vector containing the regex pattern.
}
\description{
Compile the regex used to find GitHub-closing issues in commit messages
}
\keyword{internal}
