% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GlueEscaped}
\alias{GlueEscaped}
\title{Glue without collisions of curly braces}
\usage{
GlueEscaped(
  ...,
  .sep = "\\n\\n",
  .open = "qcthatopen{",
  .close = "}qcthatclose",
  .envir = rlang::caller_env()
)
}
\arguments{
\item{...}{Values to glue and/or additional arguments passed to
\code{\link[glue:glue]{glue::glue()}}.}

\item{.sep}{(\verb{length-1 character}) Separator to use between glued values.}

\item{.open}{(\verb{length-1 character}) Opening delimiter. Defaults value avoids
collisions.}

\item{.close}{(\verb{length-1 character}) Closing delimiter. Defaults value avoids
collisions.}

\item{.envir}{(\code{environment}) Environment to evaluate expressions in.}
}
\value{
The glued expression as a length-1 character vector.
}
\description{
Glue without collisions of curly braces
}
\keyword{internal}
