% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchGQL.R
\name{GQLWrapper}
\alias{GQLWrapper}
\title{Wrap a GraphQL query with repository info}
\usage{
GQLWrapper(strQuery, strOwner, strRepo)
}
\arguments{
\item{strQuery}{(\verb{length-1 character}) The GraphQL query (or sub-query).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}
}
\value{
A single character string containing the full query.
}
\description{
Wrap a GraphQL query with repository info
}
\keyword{internal}
