% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%|0|\%}
\alias{\%|0|\%}
\title{Default value for length-0}
\usage{
x \%|0|\% y
}
\arguments{
\item{x}{An object to return if it has length > 0.}

\item{y}{An object to return if \code{x} has length 0.}
}
\value{
\code{x} if it has length > 0, otherwise \code{y}.
}
\description{
Default value for length-0
}
\keyword{internal}
