% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-default operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Object to test.}

\item{y}{Object to return if \code{x} is \code{NULL}.}
}
\value{
\code{x} or \code{y}.
}
\description{
Null-default operator
}
\keyword{internal}
