% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessIssueNumber.R
\name{GuessIssueNumber}
\alias{GuessIssueNumber}
\title{Guess the relevant issue number from the GitHub event}
\usage{
GuessIssueNumber(lGHEventPayload = LoadGHEventPayload())
}
\arguments{
\item{lGHEventPayload}{(\code{list}) The GitHub event payload. Defaults to the
result of \code{\link[=LoadGHEventPayload]{LoadGHEventPayload()}}.}
}
\value{
An integer representing the issue number (if the event payload
contains \code{issue$number} or \code{inputs$issueNumber}), or \code{NULL} if the event is
not an \code{"issues"} event or the issue number cannot be found.
}
\description{
Determine the issue number associated with the current GitHub event, if the
workflow was triggered by an \code{"issues"} event.
}
