% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessMilestones.R
\name{GuessMilestones}
\alias{GuessMilestones}
\title{Guess relevant milestone names from the GitHub event}
\usage{
GuessMilestones(lGHEventPayload = LoadGHEventPayload())
}
\arguments{
\item{lGHEventPayload}{(\code{list}) The GitHub event payload. Defaults to the
result of \code{\link[=LoadGHEventPayload]{LoadGHEventPayload()}}.}
}
\value{
A character vector of milestone names (if the event payload contains
\code{pull_request$milestone$title}, \code{release$name}, \code{release$tag_name},
\code{inputs$milestone}, or \code{inputs$tag}), or \code{NULL} if no milestone names can
be found.
}
\description{
Determine the name of any milestones associated with the current GitHub
event, if the workflow was triggered by a \code{"pull_request"} or \code{"release"}
event, or by a \code{"workflow_dispatch"} event with \code{"milestone"} input.
}
