% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{GuessPRNumber}
\alias{GuessPRNumber}
\title{Guess the most relevant pull request number}
\usage{
GuessPRNumber(
  strPkgRoot = ".",
  strOwner = gh::gh_tree_remote(strPkgRoot)[["username"]],
  strRepo = gh::gh_tree_remote(strPkgRoot)[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The latest pull request number as an integer, or \code{integer(0)} if no
pull requests are found.
}
\description{
Tries to find a pull request associated with the active branch. If it fails,
it falls back to finding the latest pull request number, optionally filtered
by state.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  GuessPRNumber()
\dontshow{\}) # examplesIf}
}
