% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessReleaseID.R
\name{GuessReleaseID}
\alias{GuessReleaseID}
\title{Guess the relevant release id from the GitHub event}
\usage{
GuessReleaseID(
  lGHEventPayload = LoadGHEventPayload(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{lGHEventPayload}{(\code{list}) The GitHub event payload. Defaults to the
result of \code{\link[=LoadGHEventPayload]{LoadGHEventPayload()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A string representing the release id (if the event payload contains
\code{release$id}, \code{release$tag_name}, or \code{inputs$tag}), or \code{NULL} if the event
is not a \code{"release"} event or the release id cannot be found.
}
\description{
Determine the release id associated with the current GitHub event, if the
workflow was triggered by a \code{"release"} event, or by a \code{"workflow_dispatch"}
event with \code{"tag"} input
}
