% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HaveString}
\alias{HaveString}
\title{Detect strings in lists of characters}
\usage{
HaveString(lCharacters, strTarget)
}
\arguments{
\item{lCharacters}{(\code{list} of \code{character}) A list of character vectors.}

\item{strTarget}{(\verb{length-1 character}) The target string to search for.}
}
\value{
A logical vector indicating whether \code{strTarget} is present in each
element of \code{lCharacters}.
}
\description{
Detect strings in lists of characters
}
\keyword{internal}
