% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_github_actions.R
\name{init_github_actions}
\alias{init_github_actions}
\title{Initialize GitHub Actions YAML File for Reporting}
\usage{
init_github_actions()
}
\value{
\code{.github/workflows} folder at root directory; \code{qualification_report.yaml} file.
}
\description{
Initialize GitHub Actions YAML File for Reporting
}
