% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_report.R
\name{init_report}
\alias{init_report}
\title{Initialize Qualification Report Template}
\usage{
init_report(package_name = read.dcf("DESCRIPTION")[[1]])
}
\description{
Initialize Qualification Report Template
}
\keyword{internal}
