% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_specs.R
\name{init_specs}
\alias{init_specs}
\title{Initialize Specifications Directory and Template File}
\usage{
init_specs()
}
\value{
\code{qualification} folder in \code{inst}; \code{qualification_specs.csv} template file in \code{inst/qualification}
}
\description{
Initialize Specifications Directory and Template File
}
\examples{
\dontrun{
init_specs()
}

}
