% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_test_cases.R
\name{init_test_cases}
\alias{init_test_cases}
\title{Initialize testing directory structure and template file.}
\usage{
init_test_cases()
}
\value{
\code{tests/testqualification/qualification} directory; \code{test_qual_T1_1.R} file.
}
\description{
Initialize testing directory structure and template file.
}
