% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize \code{{qcthat}} testing framework.}
\usage{
initialize()
}
\value{
Testing framework described in \code{{qcthat}} README.
}
\description{
Initialize \code{{qcthat}} testing framework.
}
\examples{
\dontrun{
initialize()
}

}
