% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{InstallAction}
\alias{InstallAction}
\title{Install an action from qcthat into a package repo}
\usage{
InstallAction(
  strActionName,
  lglOverwrite = FALSE,
  strPkgRoot = ".",
  envCall = rlang::caller_env()
)
}
\arguments{
\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
The path to the created GitHub Action YAML file (invisibly).
}
\description{
Install an action from qcthat into a package repo
}
\keyword{internal}
