% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{InstallFile}
\alias{InstallFile}
\title{Install a file from qcthat into a package repo}
\usage{
InstallFile(
  chrSourcePath,
  chrTargetPath,
  strExtension,
  lglOverwrite = FALSE,
  strPkgRoot = ".",
  envCall = rlang::caller_env()
)
}
\arguments{
\item{chrSourcePath}{(\code{character}) Components of a path to a source file. The
file extension should be included in the filename or omitted (NOT sent as a
separate piece of the vector).}

\item{chrTargetPath}{(\code{character}) Components of a path to a target file. The
file extension should be included in the filename or omitted (NOT sent as a
separate piece of the vector).}

\item{strExtension}{(\verb{length-1 character}) The file extension to use for the
target file. If the target path already includes an extension, it will be
replaced with this value. If the value is already correct, this won't have
any effect.}

\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
The path to the created file (invisibly).
}
\description{
Install a file from qcthat into a package repo
}
\keyword{internal}
