% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skills.R
\name{InstallSkill}
\alias{InstallSkill}
\title{Install a skill from qcthat into a package repo}
\usage{
InstallSkill(
  strSkillName,
  strTargetDir = ".github/skills",
  lglOverwrite = FALSE,
  strPkgRoot = ".",
  envCall = rlang::caller_env()
)
}
\arguments{
\item{strTargetDir}{(\verb{length-1 character}) Path to the directory where the
file should be added.}

\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_find]{gert::git_find()}}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
The path to the created skill file (invisibly).
}
\description{
Install a skill from qcthat into a package repo
}
\keyword{internal}
