% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{IsCheckingUAT}
\alias{IsCheckingUAT}
\title{Detect whether the user is specifically checking UAT issues}
\usage{
IsCheckingUAT(strUATEnvVar = "qcthat_UAT")
}
\arguments{
\item{strUATEnvVar}{(\verb{length-1 character}) The name of the environment
variable to check.}
}
\value{
\code{TRUE} if the specified environment variable is set to \code{"TRUE"}
(case-insensitive), \code{FALSE} otherwise.
}
\description{
Checks the value of an environment variable (default: \code{qcthat_UAT}) to
determine if the user is intentionally checking user-acceptance tests.
}
\examples{
CurrentValue <- Sys.getenv("qcthat_UAT")
Sys.setenv(qcthat_UAT = "")
IsCheckingUAT()
Sys.setenv(qcthat_UAT = "true")
IsCheckingUAT()
Sys.setenv(qcthat_UAT = CurrentValue)
}
