% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssueClosers.R
\name{IsIssueCloserFromRepo}
\alias{IsIssueCloserFromRepo}
\title{Check whether an issue closer originates from the target repository}
\usage{
IsIssueCloserFromRepo(lIssueCloser, strNameWithOwner)
}
\arguments{
\item{lIssueCloser}{(\code{list}) A single raw issue node.}

\item{strNameWithOwner}{(\code{character(1)}) \code{"owner/repo"} of the target repo.}
}
\value{
A length-1 \code{logical}.
}
\description{
Check whether an issue closer originates from the target repository
}
\keyword{internal}
