% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{IsOnline}
\alias{IsOnline}
\title{Check whether the user is online}
\usage{
IsOnline()
}
\value{
\code{TRUE} if the user is online.
}
\description{
Check whether the user is online
}
\keyword{internal}
