% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssueClosers.R
\name{IsTimelineNodeFromRepo}
\alias{IsTimelineNodeFromRepo}
\title{Check whether a single timeline node originates from the target repository}
\usage{
IsTimelineNodeFromRepo(lNode, strNameWithOwner)
}
\arguments{
\item{lNode}{(\code{list}) A single timeline item node.}

\item{strNameWithOwner}{(\code{character(1)}) \code{"owner/repo"} of the target repo.}
}
\value{
A length-1 \code{logical}.
}
\description{
Check whether a single timeline node originates from the target repository
}
\keyword{internal}
