% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{JoinLineCommits}
\alias{JoinLineCommits}
\title{Add blames to expanded dfFileTests}
\usage{
JoinLineCommits(dfFileTestsExpanded, dfBlames)
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the same structure as \code{dfFileTests} plus a
\code{Commits} list column containing character vectors of unique commit SHAs
that touched each test, in the order they first appear.
}
\description{
Add blames to expanded dfFileTests
}
\keyword{internal}
