% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{LinkSessionInfo}
\alias{LinkSessionInfo}
\title{Construct the URL for the session info step of a GitHub Actions job}
\usage{
LinkSessionInfo(
  strRunID = character(),
  strJobID = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run.}

\item{strJobID}{(\verb{length-1 character}) ID (typically numeric but can be very
long) of a GitHub Actions workflow run job.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A string containing the URL for the session info step (if any) in
the specified GitHub Actions job.
}
\description{
Construct the URL for the session info step of a GitHub Actions job
}
\keyword{internal}
