% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTestsFromFiles.R
\name{ListTestFiles}
\alias{ListTestFiles}
\title{Find test files}
\usage{
ListTestFiles(strTestDir)
}
\arguments{
\item{strTestDir}{(\verb{length-1 character}) Path to the directory containing
test files. Defaults to \code{"tests/testthat"}.}
}
\value{
Same as \code{\link[testthat:find_test_scripts]{testthat::find_test_scripts()}}. This function exists for
easier mocking during tests.
}
\description{
Find test files
}
\keyword{internal}
