% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGHEventPayload.R
\name{LoadGHEventPayload}
\alias{LoadGHEventPayload}
\title{Load the GitHub event payload}
\usage{
LoadGHEventPayload(strGHEventPath = Sys.getenv("GITHUB_EVENT_PATH"))
}
\arguments{
\item{strGHEventPath}{(\verb{length-1 character}) Path to the GitHub event payload
JSON file. Defaults to the \code{GITHUB_EVENT_PATH} environment variable.}
}
\value{
A list containing the parsed JSON payload, or \code{NULL} if
\code{strGHEventPath} is empty or does not point to a GitHub event payload.
}
\description{
Read the JSON file located at \code{GITHUB_EVENT_PATH} and confirm that it
resembles a GitHub event payload.
}
