% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{LoadPkgEnv}
\alias{LoadPkgEnv}
\title{Load the package environment for coverage tracing}
\usage{
LoadPkgEnv(strTestDir, envCall = rlang::caller_env())
}
\arguments{
\item{strTestDir}{(\verb{length-1 character}) Path to the directory containing
test files. Defaults to \code{"tests/testthat"}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
The loaded package environment.
}
\description{
Thin wrapper around \code{\link[pkgload:load_all]{pkgload::load_all()}} so it can be mocked in tests.
}
\keyword{internal}
