% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveUATIssues.R
\name{LoadUATIssues}
\alias{LoadUATIssues}
\title{Load UAT issues from disk}
\usage{
LoadUATIssues(strPath = "UATIssues.rds")
}
\arguments{
\item{strPath}{(\verb{length-1 character}) Path to the UAT file. Defaults to
\code{"UATIssues.rds"} in the working directory.}
}
\value{
The UAT issues data frame, invisibly.
}
\description{
Load a registry of user acceptance testing (UAT) issues from an RDS file into
the internal package environment.
}
