% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{LogUAT}
\alias{LogUAT}
\title{Log ExpectUserAccepts results}
\usage{
LogUAT(
  intParentIssue,
  intUATIssue,
  strDescription,
  strDisposition,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  dttmTimestamp = Sys.time()
)
}
\arguments{
\item{intParentIssue}{(\verb{length-1 integer}) The number of the parent issue in
a parent-child issue relationship.}

\item{intUATIssue}{(\code{integer}) The number of an issue that tracks
user-acceptance testing.}

\item{strDescription}{(\verb{length-1 character}) A brief description of a user
expectation.}

\item{strDisposition}{(\verb{length-1 character}) The result of a test, such as
"pass", "fail", or "skip", or "accepted" or "pending" for UAT.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{dttmTimestamp}{(\code{POSIXct}) A system timestamp.}
}
\value{
The value of \code{envQcthat$UATIssues}, invisibly
}
\description{
Log ExpectUserAccepts results
}
\keyword{internal}
