% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoPRs.R
\name{LookupPRFromList}
\alias{LookupPRFromList}
\title{Look up a PR from a list by number}
\usage{
LookupPRFromList(lPRs, intPRNumber, envCall = rlang::caller_env())
}
\arguments{
\item{lPRs}{A list of raw pull request objects as returned by
\code{\link[=FetchRawRepoPRs]{FetchRawRepoPRs()}}.}

\item{intPRNumber}{(\verb{length-1 integer}) The pull request number to look up.}

\item{envCall}{(\code{environment}) The calling environment for error messages.}
}
\value{
A list representing a raw pull request object.
}
\description{
Look up a PR from a list by number
}
\keyword{internal}
