% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{MakeITRCoverageFooter}
\alias{MakeITRCoverageFooter}
\title{Add a coverage summary message to ITR footer}
\usage{
MakeITRCoverageFooter(
  intIssues,
  chrTests,
  lglUseEmoji = getOption("qcthat.emoji", TRUE)
)
}
\arguments{
\item{intIssues}{(\code{integer}) A vector of issue numbers from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix or from GitHub.}

\item{chrTests}{(\code{character}) A vector of test descriptions from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
One or two string summary messages or \code{NULL} if neither message
applies.
}
\description{
Add a coverage summary message to ITR footer
}
\keyword{internal}
