% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintIssueTestMatrix.R
\name{MakeITRIgnoredLabelsFooter}
\alias{MakeITRIgnoredLabelsFooter}
\title{Add an ignored issue summary message to ITR footer}
\usage{
MakeITRIgnoredLabelsFooter(
  lIgnoredIssues,
  lglUseEmoji = getOption("qcthat.emoji", TRUE),
  lglShowIgnoredLabels = TRUE
)
}
\arguments{
\item{lIgnoredIssues}{(\code{list}) A named list of integer vectors, where each
name is an ignored label and each integer vector contains the issue numbers
that were ignored for that label.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}

\item{lglShowIgnoredLabels}{(\verb{length-1 logical}) Whether to show information
in reports about issue labels (such as \code{"qcthat-nocov"}) that have been
ignored.}
}
\value{
A string summary messages or \code{NULL} if no ignored issues are
present.
}
\description{
Add an ignored issue summary message to ITR footer
}
\keyword{internal}
