% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{MakeKeyItem}
\alias{MakeKeyItem}
\title{Helper to insert emojis into key items}
\usage{
MakeKeyItem(strCondition, lglUseEmoji = getOption("qcthat.emoji", TRUE))
}
\arguments{
\item{strCondition}{(\verb{length-1 character}) The condition to create a key item
for.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE}) or
ASCII.}
}
\value{
A length-1 character vector representing the key item with an emoji.
}
\description{
Helper to insert emojis into key items
}
\keyword{internal}
