% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRepoIssuesToCommits.R
\name{MapIssueClosersToCommits}
\alias{MapIssueClosersToCommits}
\title{Add Commits list column to dfIssueClosers}
\usage{
MapIssueClosersToCommits(
  dfIssueClosers,
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  lPRs = NULL
)
}
\arguments{
\item{dfIssueClosers}{(\code{data.frame}) The \code{\link[tibble:tibble]{tibble::tibble()}} returned by
\code{\link[=FetchRepoIssueClosers]{FetchRepoIssueClosers()}}.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}

\item{lPRs}{(\code{list} or \code{NULL}) Optional list of raw pull request objects as
returned by \code{\link[=FetchRawRepoPRs]{FetchRawRepoPRs()}}. If provided, PRs will be looked up from
this list instead of fetching individually from the API.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Issue}: Issue number (integer).
\item \code{Commits}: List column containing character vectors of commit SHAs
associated with each issue.
}
}
\description{
Add Commits list column to dfIssueClosers
}
