% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRepoIssuesToCommits.R
\name{MapRepoIssuesToCommits}
\alias{MapRepoIssuesToCommits}
\title{Map repository issues to commits}
\usage{
MapRepoIssuesToCommits(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Issue}: Issue number (integer).
\item \code{Commits}: List column containing character vectors of commit SHAs
associated with each issue.
}
}
\description{
Fetch all closed issues in a repository and expand their closers (commits or
pull requests) into the set of commits that might be related to each issue.
}
\keyword{internal}
