% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{MapTestsToCommits}
\alias{MapTestsToCommits}
\title{Map tests to commits}
\usage{
MapTestsToCommits(dfFileTests, envCall = rlang::caller_env())
}
\arguments{
\item{dfFileTests}{(\code{data.frame}) A \code{\link[tibble:tibble]{tibble::tibble()}} with the structure
returned by \code{\link[=ExtractTestsFromFiles]{ExtractTestsFromFiles()}}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the same structure as \code{dfFileTests} plus a
\code{Commits} list column containing character vectors of unique commit SHAs
that touched each test, in the order they first appear.
}
\description{
Add a column of commit SHAs to a data frame of tests by using git blame to
identify which commits last modified each line of each test.
}
\keyword{internal}
