% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{MaybeRerunAllQCPRWorkflows}
\alias{MaybeRerunAllQCPRWorkflows}
\title{Rerun QCPR workflows for PRs if UAT conditions are met}
\usage{
MaybeRerunAllQCPRWorkflows(
  dfOpenPRRefs,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{dfOpenPRRefs}{(\code{data.frame}) A tibble of PR references, as returned by
\code{FetchAllIssuePRRefs()}. Must contain columns \code{PR}, \code{HeadRef}, and \code{SHA}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
\code{dfOpenPRRefs} (invisibly). Called for side effects.
}
\description{
Rerun QCPR workflows for PRs if UAT conditions are met
}
\keyword{internal}
