% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{MaybeRerunQCPRWorkflow}
\alias{MaybeRerunQCPRWorkflow}
\title{Rerun the QCPR workflow for a specific PR and commit}
\usage{
MaybeRerunQCPRWorkflow(
  intPRNumber,
  strPRHeadRef,
  strCommitSHA,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about.}

\item{strPRHeadRef}{(\verb{length-1 character}) The branch name (head ref) of the
PR.}

\item{strCommitSHA}{(\verb{length-1 character}) The commit SHA to target.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
\code{NULL} (invisibly). Called for side effects.
}
\description{
Rerun the QCPR workflow for a specific PR and commit
}
\keyword{internal}
