% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{NormalizeDescriptionPrefix}
\alias{NormalizeDescriptionPrefix}
\title{Normalize the label description prefix}
\usage{
NormalizeDescriptionPrefix(strLabelDescription)
}
\arguments{
\item{strLabelDescription}{(\verb{length-1 character}) The description for the
label.}
}
\value{
The label description with the prefix \code{"{qcthat}: "}.
}
\description{
Normalize the label description prefix
}
\keyword{internal}
