% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{NormalizeLabelPrefix}
\alias{NormalizeLabelPrefix}
\title{Normalize the label prefix}
\usage{
NormalizeLabelPrefix(strLabel)
}
\arguments{
\item{strLabel}{(\verb{length-1 character}) The name of the label to create.}
}
\value{
The label with the prefix \code{"qcthat-"}.
}
\description{
Normalize the label prefix
}
\keyword{internal}
