% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{NullIfEmpty}
\alias{NullIfEmpty}
\title{Flatten empty vectors into NULL}
\usage{
NullIfEmpty(x)
}
\arguments{
\item{x}{An object to potentially flatten.}
}
\value{
\code{NULL} if \code{x} has length 0, otherwise \code{x}.
}
\description{
Flatten empty vectors into NULL
}
\keyword{internal}
