% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{OnCran}
\alias{OnCran}
\title{Check whether this is being tested on CRAN}
\usage{
OnCran()
}
\value{
Logical indicating whether this is running on CRAN (according to
testthat).
}
\description{
Check whether this is being tested on CRAN
}
\keyword{internal}
