% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{PrepareGQLQuery}
\alias{PrepareGQLQuery}
\title{Prepare a GraphQL query string}
\usage{
PrepareGQLQuery(
  ...,
  strOpen = "<",
  strClose = ">",
  envCall = rlang::caller_env()
)
}
\arguments{
\item{...}{(\code{character}) Lines of the query.}

\item{strOpen}{(\verb{length-1 character}) The opening delimiter for
\code{\link[glue:glue]{glue::glue()}}.}

\item{strClose}{(\verb{length-1 character}) The closing delimiter for
\code{\link[glue:glue]{glue::glue()}}.}
}
\value{
A single character string containing the formatted query.
}
\description{
Prepare a GraphQL query string
}
\keyword{internal}
