% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTestIssueContext.R
\name{PrepareTestIssueContext}
\alias{PrepareTestIssueContext}
\title{Prepare test issue context for analysis}
\usage{
PrepareTestIssueContext(
  dfPotentialIssues = NULL,
  strTestDir = "tests/testthat",
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{dfPotentialIssues}{(\code{tibble}) A data frame as returned by
\code{\link[=MapTestFilesToPotentialIssues]{MapTestFilesToPotentialIssues()}}.}

\item{strTestDir}{(\verb{length-1 character}) Path to the directory containing
test files. Defaults to \code{"tests/testthat"}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Test}: Test description (character).
\item \code{File}: Test file name (character).
\item \code{LineStart}: Starting line number of test.
\item \code{LineEnd}: Ending line number of test.
\item \code{Issues}: List column containing integer vectors of issue numbers already
tagged in the test description.
\item \code{PotentialIssueDetails}: List column of tibbles with enriched issue
details.
\item \code{TestCode}: List column of character vectors containing test code.
}
}
\description{
Collect comprehensive information about tests and their potential related
issues using \code{\link[=MapTestFilesToPotentialIssues]{MapTestFilesToPotentialIssues()}}, and enrich with test code and
enriched issue details from GitHub.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  PrepareTestIssueContext()
\dontshow{\}) # examplesIf}
}
