% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PrettyTimestamp}
\alias{PrettyTimestamp}
\title{Get the formatted time}
\usage{
PrettyTimestamp(strTimezone = "UTC")
}
\arguments{
\item{strTimezone}{(\verb{length-1 character}) Timezone to use for formatting.}
}
\value{
A length-1 character vector representing the current time in the
given timezone.
}
\description{
Get the formatted time
}
\keyword{internal}
