% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{printing}
\alias{printing}
\alias{print.qcthat_Object}
\alias{format.qcthat_Object}
\title{Printing qcthat objects}
\usage{
\method{print}{qcthat_Object}(x, ...)

\method{format}{qcthat_Object}(
  x,
  ...,
  lglUseEmoji = getOption("qcthat.emoji", TRUE),
  transform = identity
)
}
\arguments{
\item{x}{(\code{qcthat_Object}) The qcthat object to format.}

\item{...}{Additional arguments passed to methods.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}

\item{transform}{(\code{function}) A function to transform the output before
returning it. By default, this is \code{\link[base:identity]{base::identity()}} for \code{format()}
(returns the formatted output), and \code{\link[base:writeLines]{base::writeLines()}} for \code{print()}
(sends the output to \code{\link[=stdout]{stdout()}}).}
}
\description{
One of the main purposes of qcthat is to produce user-friendly reports about
a package's issue and test statuses. Printing an issue-test matrix provides a
nested view of the relationship between milestones, issues, and tests, as
well as the state of issues ("open", "closed (completed)", or "closed (won't
fix)") and the disposition of tests ("passed", "failed", or "skipped").
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  lTestResults <- testthat::test_local(
    stop_on_failure = FALSE,
    reporter = "silent"
  )
  IssueTestMatrix <- CompileIssueTestMatrix(
    dfRepoIssues = FetchRepoIssues(),
    dfTestResults = CompileTestResults(lTestResults)
  )
  print(IssueTestMatrix)

  # Remove the "qcthat_IssueTestMatrix" class to see the raw tibble.
  print(tibble::as_tibble(IssueTestMatrix))
\dontshow{\}) # examplesIf}
}
