% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{QCMergeGH}
\alias{QCMergeGH}
\title{Generate a QC report of issues associated with a GitHub merge}
\usage{
QCMergeGH(
  strSourceRef = GetActiveBranch(strPkgRoot),
  strTargetRef = GetDefaultBranch(strPkgRoot),
  intPageMax = 100L,
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels(),
  dfITM = NULL,
  envCall = rlang::caller_env()
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{intPageMax}{(\verb{length-1 integer}) The maximum number of pages of commits
to fetch from the GitHub API. Each page contains up to 100 commits.
Defaults to 100, which fetches up to 10,000 commits. You likely never need
to increase this number, but try a larger number if a merge involves a very
large number of commits in a very large repository.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}} (often via \code{\link[=QCPackage]{QCPackage()}}).}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to issues that are associated with pull requests that will be
merged when \code{strSourceRef} is merged into \code{strTargetRef}.
}
\description{
Finds all commits in \code{strSourceRef} that are not in \code{strTargetRef}, finds all
pull requests associated with those commits, finds all issues associated with
those pull requests (according to GitHub's graph of connections between
issues and commits), and generates a QC report for those issues. This is a
more robust check than \code{\link[=QCMergeLocal]{QCMergeLocal()}}. Note: If the comparison involves
more than 5000 commits, increase \code{intPageMax} to fetch additional commits in
batches of 100.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # This will only make sense if you are working in a git repository and have
  # an active branch that is different from the default branch.

  QCMergeGH()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=QCMergeLocal]{QCMergeLocal()}} to use local git data to guess connections between
issues and the commits that closed them.
}
