% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReports.R
\name{QCMilestones}
\alias{QCMilestones}
\title{Generate a QC report of a specific milestone or milestones}
\usage{
QCMilestones(
  chrMilestones,
  strPkgRoot = ".",
  strOwner = gh::gh_tree_remote(strPkgRoot)[["username"]],
  strRepo = gh::gh_tree_remote(strPkgRoot)[["repo"]],
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{chrMilestones}{(\code{character}) The name(s) of milestone(s) to filter
issues by.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to the indicated issues.
}
\description{
Generate a report about the test status of issues in a milestone or
milestones.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # This will only make sense if you are working in a git repository that has
  # a milestone named "v0.1.0" on GitHub.
  QCMilestones("v0.1.0")
\dontshow{\}) # examplesIf}
}
