% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCPR.R
\name{QCPR}
\alias{QCPR}
\title{Generate a QC report of issues associated with a GitHub pull request}
\usage{
QCPR(
  intPRNumber = GuessPRNumber(strPkgRoot, strOwner, strRepo, strGHToken),
  intPageMax = 100L,
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels(),
  dfITM = NULL,
  envCall = rlang::caller_env()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{intPageMax}{(\verb{length-1 integer}) The maximum number of pages of commits
to fetch from the GitHub API. Each page contains up to 100 commits.
Defaults to 100, which fetches up to 10,000 commits. You likely never need
to increase this number, but try a larger number if a merge involves a very
large number of commits in a very large repository.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}} (often via \code{\link[=QCPackage]{QCPackage()}}).}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to issues that will be closed by merging the pull request.
}
\description{
Find issues associated with a GitHub pull request, whether they were added
via keywords, using the pull request sidebar, or using the issue sidebar. See
\href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue}{GitHub Docs: Link a pull request to an issue}
for details on how issues can become associated with a pull request.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # You must have at least one pull request open in the GitHub repository
  # associated with the current git repository for this to return any
  # results.

  QCPR()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=QCMergeLocal]{QCMergeLocal()}} to use local git data to guess connections between
issues and the commits that closed them.
}
