% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcthat-package.R
\docType{package}
\name{qcthat-package}
\alias{qcthat}
\alias{qcthat-package}
\title{qcthat: Lightweight R Package Quality Control}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A lightweight quality control framework for R packages, particularly those used in regulated environments such as clinical trials. Integrates formal quality control reports into package development workflows using 'testthat' and 'GitHub'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilead-public.github.io/qcthat/}
  \item \url{https://github.com/Gilead-Public/qcthat}
  \item Report bugs at \url{https://github.com/Gilead-Public/qcthat/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Jeremy Wildfire \email{jwildfire@gmail.com}
  \item Jon Harmon \email{jon.harmon@atorusresearch.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
  \item Natalia Andriychuk \email{Natalia.Andriychuk@gilead.com}
  \item Matt Roumaya \email{Matt.Roumaya@atorusresearch.com} (\href{https://orcid.org/0000-0002-8009-2771}{ORCID})
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
