% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcthat.R
\name{qcthat}
\alias{qcthat}
\title{Initialize \code{{qcthat}} testing framework.}
\usage{
qcthat(package_name = read.dcf("DESCRIPTION")[[1]])
}
\arguments{
\item{package_name}{Name of package.}
}
\value{
Testing framework described in \code{{qcthat}} README.
}
\description{
Initialize \code{{qcthat}} testing framework.
}
\examples{
\dontrun{
qcthat()
}

}
