% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{qcthatAbort}
\alias{qcthatAbort}
\title{Generate a classed qcthat error}
\usage{
qcthatAbort(
  strErrorMessage,
  strErrorSubclass,
  envCall = rlang::caller_env(),
  envErrorMessage = rlang::caller_env(),
  cndParent = NULL,
  ...
)
}
\arguments{
\item{strErrorMessage}{(\verb{length-1 character}) A message to include in and
error report. Can include \code{\link[glue:glue]{glue::glue()}} syntax.}

\item{strErrorSubclass}{(\verb{length-1 character}) A subclass for an error
condition.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}

\item{envErrorMessage}{(\code{environment}) The environment to pass to
\code{\link[glue:glue]{glue::glue()}} to resolve variables.}

\item{cndParent}{(\code{condition} or \code{NULL}) Parent condition, if any.}

\item{...}{Additional parameters to pass to \code{\link[cli:cli_abort]{cli::cli_abort()}}.}
}
\description{
Generate a classed qcthat error
}
\keyword{internal}
