% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{qcthatPath}
\alias{qcthatPath}
\title{Find a path within qcthat}
\usage{
qcthatPath(chrPath)
}
\arguments{
\item{chrPath}{(\code{character}) Components of a path}
}
\value{
The full path to the file within the \code{qcthat} package.
}
\description{
Exists primarily to make it easier to mock this for testing.
}
\keyword{internal}
