% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTestCode.R
\name{ReadFileLines}
\alias{ReadFileLines}
\title{Read lines from a file}
\usage{
ReadFileLines(
  strFile,
  strDirPath,
  intLineStart,
  intLineEnd,
  envCall = rlang::caller_env()
)
}
\arguments{
\item{strFile}{(\verb{length-1 character}) A file name without the path.}

\item{strDirPath}{(\verb{length-1 character}) The path to a directory.}

\item{intLineStart}{(\verb{length-1 integer}) The starting line number.}

\item{intLineEnd}{(\verb{length-1 integer}) The ending line number.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
A \code{character} vector containing a subset of lines from the file.
}
\description{
Read lines from a file
}
\keyword{internal}
