% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveAsJSON.R
\name{ReadJSONAsIssueTestMatrix}
\alias{ReadJSONAsIssueTestMatrix}
\title{Read a JSON file as an IssueTestMatrix}
\usage{
ReadJSONAsIssueTestMatrix(strJSONPath)
}
\arguments{
\item{strJSONPath}{(\verb{length-1 character}) File path to save or read JSON.}
}
\value{
A \code{qcthat_IssueTestMatrix} object.
}
\description{
Read a JSON file that was saved with \code{SaveAsJSON.qcthat_IssueTestMatrix()}
and return it as a \code{qcthat_IssueTestMatrix} object.
}
