% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RecodeValues}
\alias{RecodeValues}
\title{Choose between recode functions based on dplyr version}
\usage{
RecodeValues(x, ..., default = NULL)
}
\arguments{
\item{x}{(\code{vector}) The vector to recode.}

\item{...}{Recode pairs in the form of \code{old ~ new}. See \code{\link[dplyr:case_match]{dplyr::case_match()}}
or \code{\link[dplyr:recode_values]{dplyr::recode_values()}} for details.}

\item{default}{(\code{scalar}) The default value to use for unmatched cases. See
\code{\link[dplyr:case_match]{dplyr::case_match()}} or \code{\link[dplyr:recode_values]{dplyr::recode_values()}} for details.}
}
\value{
A vector with the same size as \code{x} with values recoded according to
the specified pairs and default.
}
\description{
Choose between recode functions based on dplyr version
}
\keyword{internal}
