% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{RemovePRsFromIssues}
\alias{RemovePRsFromIssues}
\title{Get rid of PRs in the issues list}
\usage{
RemovePRsFromIssues(lIssuesRaw)
}
\arguments{
\item{lIssuesRaw}{(\code{list}) List of raw issue objects as returned by
\code{\link[gh:gh]{gh::gh()}}.}
}
\value{
A list of issue objects without PRs.
}
\description{
Get rid of PRs in the issues list
}
\keyword{internal}
