% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{RerunWorkflow}
\alias{RerunWorkflow}
\title{Trigger a rerun of a GitHub Action workflow run}
\usage{
RerunWorkflow(
  strRunID,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strRunID}{(\verb{length-1 character, double, or integer}) The ID of the
workflow run to rerun.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
An empty GitHub API return with status code 201.
}
\description{
Trigger a rerun of a GitHub Action workflow run
}
\keyword{internal}
