% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{RerunWorkflowIfFinished}
\alias{RerunWorkflowIfFinished}
\title{Rerun a workflow only if no instances are currently in progress}
\usage{
RerunWorkflowIfFinished(
  lPRActionRuns,
  strCommitSHA,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{lPRActionRuns}{(\code{list}) A list of workflow run objects as returned by
GitHub.}

\item{strCommitSHA}{(\verb{length-1 character}) The commit SHA to target.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
\code{NULL} (invisibly). Called for side effects.
}
\description{
Rerun a workflow only if no instances are currently in progress
}
\keyword{internal}
