% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveAsJSON.R
\name{SaveAsJSON}
\alias{SaveAsJSON}
\alias{SaveAsJSON.default}
\alias{SaveAsJSON.qcthat_IssueTestMatrix}
\title{Save an object as JSON}
\usage{
SaveAsJSON(x, strJSONPath)

\method{SaveAsJSON}{default}(x, strJSONPath)

\method{SaveAsJSON}{qcthat_IssueTestMatrix}(x, strJSONPath)
}
\arguments{
\item{x}{(\code{any}) An R object to save as JSON.}

\item{strJSONPath}{(\verb{length-1 character}) File path to save or read JSON.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Save an object as formatted JSON. This is a generic function with a method
for \code{qcthat_IssueTestMatrix} objects that preserves attributes and column
order.
}
