% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveUATIssues.R
\name{SaveUATIssues}
\alias{SaveUATIssues}
\title{Save UAT issues to disk}
\usage{
SaveUATIssues(strPath = "UATIssues.rds")
}
\arguments{
\item{strPath}{(\verb{length-1 character}) Path to the UAT file. Defaults to
\code{"UATIssues.rds"} in the working directory.}
}
\value{
\code{NULL} invisibly.
}
\description{
Save the internal registry of user acceptance testing (UAT) issues to an RDS
file. This allows the UAT state to be persisted across R sessions.
}
