% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{SeparateParentColumn}
\alias{SeparateParentColumn}
\title{Parse the ParentUrl column into its components}
\usage{
SeparateParentColumn(dfIssues)
}
\arguments{
\item{dfIssues}{(\code{data.frame}) Data frame with a \code{ParentUrl} column.}
}
\value{
The input data frame with \code{ParentUrl} split into \code{ParentOwner},
\code{ParentRepo}, and \code{ParentNumber} columns.
}
\description{
Parse the ParentUrl column into its components
}
\keyword{internal}
