% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{SetupGHLabels}
\alias{SetupGHLabels}
\title{Setup qcthat labels in a GitHub repository}
\usage{
SetupGHLabels(
  dfLabels = DefaultIgnoreLabelsDF(),
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{dfLabels}{(\code{data.frame}) A data frame with columns \code{Label},
\code{Description}, and \code{Color}, specifying the labels to create. By default,
this is the data frame returned by \code{\link[=DefaultIgnoreLabelsDF]{DefaultIgnoreLabelsDF()}}. Descriptions
of labels created via this function are prefixed with \code{"{qcthat}: "} to
make it easier to search for them in your list of labels.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Create the qcthat labels in a GitHub repository if those labels do not
already exist.
}
