% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{chrClass}{(\code{character}) Class name(s) to assign to the object.}

\item{chrCommitSHAs}{(\code{character}) SHAs of git commits.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{chrKeywords}{(\code{character}) Keywords to search for just before issue
numbers in commit messages. Defaults to the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub issue-linking keywords}}

\item{chrMilestones}{(\code{character}) The name(s) of milestone(s) to filter
issues by.}

\item{chrSourcePath}{(\code{character}) Components of a path to a source file. The
file extension should be included in the filename or omitted (NOT sent as a
separate piece of the vector).}

\item{chrTargetPath}{(\code{character}) Components of a path to a target file. The
file extension should be included in the filename or omitted (NOT sent as a
separate piece of the vector).}

\item{chrTests}{(\code{character}) A vector of test descriptions from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix.}

\item{dfITM}{(\code{qcthat_IssueTestMatrix}) A \code{qcthat_IssueTestMatrix} object as
returned by \code{\link[=AsIssueTestMatrix]{AsIssueTestMatrix()}}.}

\item{dfRepoIssues}{(\code{qcthat_Issues} or data frame) Data frame of GitHub
issues as returned by \code{\link[=FetchRepoIssues]{FetchRepoIssues()}}.}

\item{dfTestResults}{(\code{qcthat_TestResults} or data frame) Data frame of test
results as returned by \code{\link[=CompileTestResults]{CompileTestResults()}}.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}

\item{fctDisposition}{(\code{factor}) Disposition factor with levels \code{c("fail", "skip", "pass")}.}

\item{intIssues}{(\code{integer}) A vector of issue numbers from a
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}} matrix or from GitHub.}

\item{intMaxCommits}{(\verb{length-1 integer}) The maximum number of commits to
return from git logs. Leaving this at the default should almost always be
fine, but you can reduce the number if your repository has a long commit
history and this function is slow.}

\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about.}

\item{intPRNumbers}{(\code{integer}) A vector of pull request numbers.}

\item{lIssuesNonPR}{(\code{list}) List of issue objects as returned by
\code{\link[=RemovePRsFromIssues]{RemovePRsFromIssues()}}.}

\item{lTestResults}{(\code{testthat_results}) A testthat test results object,
typically obtained by running something like \code{\link[testthat:test_package]{testthat::test_local()}} with
\code{stop_on_failure = FALSE} and a reporter that doesn't cause issues in
parallel testing, like \code{reporter = "silent"}, and assigning it to a name.}

\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{lglShowMilestones}{(\verb{length-1 logical}) Whether to separate issues by
milestones in reports.}

\item{lglShowIgnoredLabels}{(\verb{length-1 logical}) Whether to show information
in reports about issue labels (such as \code{"qcthat-nocov"}) that have been
ignored.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{objShape}{(\verb{0-row data.frame}, etc) Object with the expected structure.}

\item{strExtension}{(\verb{length-1 character}) The file extension to use for the
target file. If the target path already includes an extension, it will be
replaced with this value. If the value is already correct, this won't have
any effect.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strState}{(\verb{length-1 character}) State of issues or pull requests to
fetch. Must be one of \code{"open"}, \code{"closed"}, or \code{"all"}. Defaults to
\code{"open"} for pull requests and \code{"all"} for issues.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}
}
\description{
Reused parameter definitions are gathered here for easier usage. Use
\verb{@inheritParams shared-params} to document parameters defined here.
}
\keyword{internal}
