% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SimplePluralize}
\alias{SimplePluralize}
\title{Add an s to a word based on count}
\usage{
SimplePluralize(strSingular, intN)
}
\arguments{
\item{strSingular}{(\code{character}) Singular form of the word.}

\item{intN}{(\code{integer}) Count to determine singular vs. plural.}
}
\value{
A character string with the singular or plural form of the word.
}
\description{
Add an s to a word based on count
}
\keyword{internal}
