% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SplitFlattenCommas}
\alias{SplitFlattenCommas}
\title{Flatten a character vector of comma-separated values}
\usage{
SplitFlattenCommas(chrVector)
}
\arguments{
\item{chrVector}{(\code{character}) The character vector to split. If any element
contains commas, it will be split, and spaces around the commas will be
removed.}
}
\value{
A character vector with the elements of \code{chrVector} split by commas
and flattened into a single vector.
}
\description{
Flatten a character vector of comma-separated values
}
\keyword{internal}
