% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTestCode.R
\name{SubsetFileLines}
\alias{SubsetFileLines}
\title{Subset lines from a file}
\usage{
SubsetFileLines(
  strFilePath,
  intLineStart,
  intLineEnd,
  envCall = rlang::caller_env()
)
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}

\item{intLineStart}{(\verb{length-1 integer}) The starting line number.}

\item{intLineEnd}{(\verb{length-1 integer}) The ending line number.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{character} vector containing a subset of lines from the file.
}
\description{
Subset lines from a file
}
\keyword{internal}
