% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{TallyCoverage}
\alias{TallyCoverage}
\title{Tally coverage (wrapper for mocking)}
\usage{
TallyCoverage(objCoverage)
}
\arguments{
\item{objCoverage}{A \code{coverage} object.}
}
\value{
A data frame as returned by \code{\link[covr:tally_coverage]{covr::tally_coverage()}}.
}
\description{
Tally coverage (wrapper for mocking)
}
\keyword{internal}
