% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchUAIssue.R
\name{TitleUAIssue}
\alias{TitleUAIssue}
\title{Generate a title for a UAT issue}
\usage{
TitleUAIssue(strDescription, intIssue)
}
\arguments{
\item{strDescription}{(\verb{length-1 character}) A brief description of a user
expectation.}

\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}
}
\value{
A string title for the UAT issue.
}
\description{
Generate a title for a UAT issue
}
\keyword{internal}
