% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{UpdateCommentRaw}
\alias{UpdateCommentRaw}
\title{Update a GitHub comment}
\usage{
UpdateCommentRaw(
  dblCommentGHID,
  strBodyCompiled,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strBodyCompiled}{(\verb{length-1 character}) The full body of an issue, PR,
or comment, in GitHub markdown, including components such as a title and a
hidden \code{qcthat-comment-id}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The comment object as returned by \code{\link[gh:gh]{gh::gh()}}, invisibly.
}
\description{
Update a GitHub comment
}
\keyword{internal}
