% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateIssue.R
\name{UpdateIssue}
\alias{UpdateIssue}
\title{Update a GitHub issue}
\usage{
UpdateIssue(
  intIssue,
  ...,
  strTitle = NULL,
  strBody = NULL,
  strState = NULL,
  strStateReason = NULL,
  strMilestone = NULL,
  chrLabels = NULL,
  chrAssignees = NULL,
  strType = NULL,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue to update.}

\item{...}{Additional arguments to pass to \code{\link[=CallGHAPI]{CallGHAPI()}}. Any \code{NULL} values
are discarded.}

\item{strTitle}{(\verb{length-1 character}) A title for an issue.}

\item{strBody}{(\verb{length-1 character}) The body of an issue, PR, comment, or
release, in GitHub markdown.}

\item{strState}{(\verb{length-1 integer}) The state to set. Must be one of
\code{"open"}, \code{"closed"}, or \code{NULL} (to ignore this field).}

\item{strStateReason}{(\verb{length-1 character}) The reason for the state change
(for \code{"closed"} \code{strState}). Must be one of \code{"completed"}, \code{"not_planned"},
\code{"duplicate"}, \code{"reopened."}, or \code{NULL}.}

\item{strMilestone}{(\verb{length-1 character}) The milestone to set. Must be the
title of an existing milestone, \code{NULL} (to ignore this field), or
\code{character()} to remove the milestone.}

\item{chrLabels}{(\code{character}) The name(s) of labels(s) to use.}

\item{chrAssignees}{(\code{character}) The assignees to set. Must be the logins of
existing users, \code{NULL} (to ignore this field), or \code{character()} to remove
all assignees.}

\item{strType}{(\verb{length-1 character}) The type of issue to set. Must be the
name of an issue type available to this repo, or \code{NULL} (to ignore this
field).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A \code{qcthat_Issues} object, which is a \code{\link[tibble:tibble]{tibble::tibble()}} with
columns:
\itemize{
\item \code{Issue}: Issue number.
\item \code{Title}: Issue title.
\item \code{Body}: Issue body (the full text of the issue).
\item \code{Labels}: List column of character vectors of issue labels.
\item \code{State}: Issue state (\code{open} or \code{closed}).
\item \code{StateReason}: Reason for issue state (e.g., \code{completed}) or \code{NA} if not
applicable.
\item \code{Milestone}: Issue milestone title or \code{NA} if not applicable.
\item \code{Type}: Issue type or \code{"Issue"} if no issue type is available.
\item \code{Url}: URL of the issue on GitHub.
\item \code{ParentOwner}: GitHub username or organization name of the parent issue
if applicable, otherwise \code{NA}.
\item \code{ParentRepo}: GitHub repository name of the parent issue if applicable,
otherwise \code{NA}.
\item \code{ParentNumber}: GitHub issue number of the parent issue if applicable,
otherwise \code{NA}.
\item \code{CreatedAt}: \code{POSIXct} timestamp of when the issue was created.
\item \code{ClosedAt}: \code{POSIXct} timestamp of when the issue was closed, or \code{NA} if
the issue is still open.
\item \code{Assignees}: List column of character vectors of GitHub usernames of
issue assignees.
}
}
\description{
Update one or more fields in a GitHub issue. \code{NULL} values are ignored, while
empty values (such as \code{character()}) often result in clearing the field. See
the \href{https://docs.github.com/en/rest/issues/issues?apiVersion=2026-03-10#update-an-issue}{GitHub API documentation}
for more details.
}
