% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{UpdateIssueComment}
\alias{UpdateIssueComment}
\title{Update or create a comment on a GitHub issue}
\usage{
UpdateIssueComment(
  intIssue,
  strBodyCompiled,
  strCommentID,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strBodyCompiled}{(\verb{length-1 character}) The full body of an issue, PR,
or comment, in GitHub markdown, including components such as a title and a
hidden \code{qcthat-comment-id}.}

\item{strCommentID}{(\verb{length-1 character}) A unique ID for a comment within a
given context, which is usually a hash of the title of the comment.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The comment object as returned by \code{\link[gh:gh]{gh::gh()}}, invisibly.
}
\description{
Update or create a comment on a GitHub issue
}
\keyword{internal}
