% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateIssue.R
\name{UpdateIssueRaw}
\alias{UpdateIssueRaw}
\title{Update an issue through the GitHub API}
\usage{
UpdateIssueRaw(
  intIssue,
  ...,
  strTitle = NULL,
  strBody = NULL,
  strState = NULL,
  strStateReason = NULL,
  strMilestone = NULL,
  chrLabels = NULL,
  chrAssignees = NULL,
  strType = NULL,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue to update.}

\item{...}{Additional arguments to pass to \code{\link[=CallGHAPI]{CallGHAPI()}}. Any \code{NULL} values
are discarded.}

\item{strTitle}{(\verb{length-1 character}) A title for an issue.}

\item{strBody}{(\verb{length-1 character}) The body of an issue, PR, comment, or
release, in GitHub markdown.}

\item{strState}{(\verb{length-1 integer}) The state to set. Must be one of
\code{"open"}, \code{"closed"}, or \code{NULL} (to ignore this field).}

\item{strStateReason}{(\verb{length-1 character}) The reason for the state change
(for \code{"closed"} \code{strState}). Must be one of \code{"completed"}, \code{"not_planned"},
\code{"duplicate"}, \code{"reopened."}, or \code{NULL}.}

\item{strMilestone}{(\verb{length-1 character}) The milestone to set. Must be the
title of an existing milestone, \code{NULL} (to ignore this field), or
\code{character()} to remove the milestone.}

\item{chrLabels}{(\code{character}) The name(s) of labels(s) to use.}

\item{chrAssignees}{(\code{character}) The assignees to set. Must be the logins of
existing users, \code{NULL} (to ignore this field), or \code{character()} to remove
all assignees.}

\item{strType}{(\verb{length-1 character}) The type of issue to set. Must be the
name of an issue type available to this repo, or \code{NULL} (to ignore this
field).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list with a raw issue object as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Update an issue through the GitHub API
}
\keyword{internal}
