% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttachReleaseReports.R
\name{UpdateReleaseBody}
\alias{UpdateReleaseBody}
\title{Update the body of a GitHub release}
\usage{
UpdateReleaseBody(
  strReleaseID,
  strBody,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strReleaseID}{(\verb{length-1 character}) ID (typically numeric but can be
very long) of a GitHub release.}

\item{strBody}{(\verb{length-1 character}) The body of an issue, PR, comment, or
release, in GitHub markdown.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The updated release object as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Update the body of a GitHub release
}
\keyword{internal}
