% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_qcthat.R
\name{use_qcthat}
\alias{use_qcthat}
\title{Set up qcthat for a package}
\usage{
use_qcthat(
  dfLabels = DefaultIgnoreLabelsDF(),
  lglOverwrite = FALSE,
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{dfLabels}{(\code{data.frame}) A data frame with columns \code{Label},
\code{Description}, and \code{Color}, specifying the labels to create. By default,
this is the data frame returned by \code{\link[=DefaultIgnoreLabelsDF]{DefaultIgnoreLabelsDF()}}. Descriptions
of labels created via this function are prefixed with \code{"{qcthat}: "} to
make it easier to search for them in your list of labels.}

\item{lglOverwrite}{(\verb{length-1 logical}) Whether to overwrite files if they
already exist.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_find]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
\code{TRUE} (invisibly).
}
\description{
Set up qcthat labels and GitHub Action workflow for a package repository.
This function combines \code{\link[=SetupGHLabels]{SetupGHLabels()}} and \code{\link[=Action_qcthat]{Action_qcthat()}} to create
the necessary GitHub labels and install the GitHub Action workflow for
managing qcthat Quality Control. We recommend reviewing the generated action
to determine whether you would like to turn any features off.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  use_qcthat()
\dontshow{\}) # examplesIf}
}
