% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{UseActionInProject}
\alias{UseActionInProject}
\title{Use a GitHub Action in a project}
\usage{
UseActionInProject(strActionFilename, strPkgRoot = ".")
}
\arguments{
\item{strActionFilename}{(\verb{length-1 character}) The filename of the GitHub
Action YAML file.}

\item{strPkgRoot}{(\verb{length-1 character}) The root directory of the package.}
}
\value{
(\verb{length-1 character}) The URL of the GitHub Action YAML file that
was added.
}
\description{
Use a GitHub Action in a project
}
\keyword{internal}
