% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gh.R
\name{UsesGit}
\alias{UsesGit}
\title{Check whether a package uses git}
\usage{
UsesGit(strPkgRoot = ".")
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}
}
\value{
A length-1 \code{logical} indicating whether the package at \code{strPkgRoot}
is a git repository.
}
\description{
Check whether a package uses git
}
\keyword{internal}
