% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{ValidateDFLabels}
\alias{ValidateDFLabels}
\title{Validate the labels data frame}
\usage{
ValidateDFLabels(dfLabels)
}
\arguments{
\item{dfLabels}{(\code{data.frame}) A data frame with columns \code{Label},
\code{Description}, and \code{Color}, specifying the labels to create. By default,
this is the data frame returned by \code{\link[=DefaultIgnoreLabelsDF]{DefaultIgnoreLabelsDF()}}. Descriptions
of labels created via this function are prefixed with \code{"{qcthat}: "} to
make it easier to search for them in your list of labels.}
}
\value{
\code{NULL} (invisibly). Errors if the data frame is invalid.
}
\description{
Validate the labels data frame
}
\keyword{internal}
