% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTestCode.R
\name{ValidateFilePath}
\alias{ValidateFilePath}
\title{Validate and construct a file path}
\usage{
ValidateFilePath(strFile, strDirPath, envCall = rlang::caller_env())
}
\arguments{
\item{strFile}{(\verb{length-1 character}) A file name without the path.}

\item{strDirPath}{(\verb{length-1 character}) The path to a directory.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{character} path to the validated file.
}
\description{
Validate and construct a file path
}
\keyword{internal}
