% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{WriteAllTempTestFiles}
\alias{WriteAllTempTestFiles}
\title{Write all temporary test files at once}
\usage{
WriteAllTempTestFiles(lstTestCode)
}
\arguments{
\item{lstTestCode}{(\code{list}) List of character vectors, one per test.}
}
\value{
A character vector of temp file paths.
}
\description{
Write all temporary test files at once
}
\keyword{internal}
