% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsExpected}
\alias{AsExpected}
\title{Assign a class and expected structure to an object}
\usage{
AsExpected(x, objShape, chrClass)
}
\arguments{
\item{x}{Object to assign the class to.}

\item{objShape}{(\verb{0-row data.frame}, etc) Object with the expected structure.}

\item{chrClass}{(\code{character}) Class name(s) to assign to the object.}
}
\value{
An object with the expected structure and class. Note: If \code{x} has a
length and/or rows, the structure is not validated against \code{objShape}.
}
\description{
Assign a class and expected structure to an object
}
\keyword{internal}
