% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{AsIssuesDF}
\alias{AsIssuesDF}
\title{Assign the qcthat_Issues class to a data frame}
\usage{
AsIssuesDF(df)
}
\arguments{
\item{df}{(\code{data.frame}) Data frame to assign the class to.}
}
\value{
A \code{qcthat_Issues} object.
}
\description{
Assign the qcthat_Issues class to a data frame
}
\keyword{internal}
