% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{AsIssueTestMatrix}
\alias{AsIssueTestMatrix}
\title{Assign the qcthat_IssueTestMatrix class to a data frame}
\usage{
AsIssueTestMatrix(df)
}
\arguments{
\item{df}{(\code{data.frame}) Data frame to assign the class to.}
}
\value{
A \code{qcthat_IssueTestMatrix} object.
}
\description{
Assign the qcthat_IssueTestMatrix class to a data frame
}
\keyword{internal}
