% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{BlameFile}
\alias{BlameFile}
\title{Git blame a file}
\usage{
BlameFile(strFilePath, envCall = rlang::caller_env())
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{File}: Path to the file, relative to the package root.
\item \code{Line}: Line number (integer).
\item \code{Commits}: List column where each element contains the commit SHA
(character) that last modified that line.
}
}
\description{
Git blame a file
}
\keyword{internal}
