% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{BlameFileRaw}
\alias{BlameFileRaw}
\title{Git blame a file (raw)}
\usage{
BlameFileRaw(strFilePath, envCall = rlang::caller_env())
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
The result of \code{\link[git2r:blame]{git2r::blame()}}. This function exists to make mocking
easier.
}
\description{
Git blame a file (raw)
}
\keyword{internal}
