% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{BuildCommitPRQuery}
\alias{BuildCommitPRQuery}
\title{Build a GraphQL sub-query for a single commit's PRs}
\usage{
BuildCommitPRQuery(strCommitSHA)
}
\arguments{
\item{strCommitSHA}{(\verb{length-1 character}) The commit SHA to target.}
}
\value{
A character string for the GraphQL sub-query.
}
\description{
Build a GraphQL sub-query for a single commit's PRs
}
\keyword{internal}
