% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{BuildCommitPRQuery}
\alias{BuildCommitPRQuery}
\title{Build a GraphQL sub-query for a single commit's PRs}
\usage{
BuildCommitPRQuery(chrSHA, intIndex)
}
\arguments{
\item{chrSHA}{(\verb{length-1 character}) The commit SHA.}

\item{intIndex}{(\verb{length-1 integer}) A unique index for the query alias.}
}
\value{
A character string for the GraphQL sub-query.
}
\description{
Build a GraphQL sub-query for a single commit's PRs
}
\keyword{internal}
