% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintSingleIssueTestResults.R
\name{ChooseStateIndicator}
\alias{ChooseStateIndicator}
\title{Choose an emoji to indicate issue state}
\usage{
ChooseStateIndicator(
  strStateReason,
  lglUseEmoji = getOption("qcthat.emoji", TRUE)
)
}
\arguments{
\item{strStateReason}{(\verb{length-1 character}) Reason for issue state (e.g.,
\code{completed}) or \code{NA} if not applicable.}

\item{lglUseEmoji}{(\verb{length-1 logical}) Whether to use emojis (if \code{TRUE} and
the emoji package is installed) or ASCII indicators (if \code{FALSE}) in the
output. By default, this is determined by the \code{qcthat.emoji} option, which
defaults to \code{TRUE}.}
}
\value{
A length-1 character vector representing the emoji or ASCII
indicator for the given condition.
}
\description{
Choose an emoji to indicate issue state
}
\keyword{internal}
