% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentIssue.R
\name{CommentIssue}
\alias{CommentIssue}
\title{Comment on a GitHub Issue}
\usage{
CommentIssue(
  intIssue,
  strTitle,
  strBody,
  strCommentID = rlang::hash(strTitle),
  lglUpdate = TRUE,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{strTitle}{(\verb{length-1 character}) A title for an issue.}

\item{strBody}{(\verb{length-1 character}) The body of an issue, PR, or comment,
in GitHub markdown.}

\item{strCommentID}{(\verb{length-1 character}) A unique ID for a comment within a
given context, which is usually a hash of the title of the comment.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The comment object as returned by \code{\link[gh:gh]{gh::gh()}}, invisibly.
}
\description{
Create or update a comment on a GitHub issue with a standardized format.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# This only works if you have an issue #1 in your repository.
CommentIssue(
  intIssue = 1,
  strTitle = "QC Report",
  strBody = "All checks passed successfully."
)
\dontshow{\}) # examplesIf}
}
