% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileIssueTestMatrix.R
\name{CompileIssueTestResultsByMilestone}
\alias{CompileIssueTestResultsByMilestone}
\title{Compile issue test results by milestone}
\usage{
CompileIssueTestResultsByMilestone(dfRepoIssues, dfTestResults)
}
\arguments{
\item{dfRepoIssues}{(\code{qcthat_Issues} or data frame) Data frame of GitHub
issues as returned by \code{\link[=FetchRepoIssues]{FetchRepoIssues()}}.}

\item{dfTestResults}{(\code{qcthat_TestResults} or data frame) Data frame of test
results as returned by \code{\link[=CompileTestResults]{CompileTestResults()}}.}
}
\value{
A data frame with \code{IssueTestResults} nested by \code{Milestone}
}
\description{
Compile issue test results by milestone
}
\keyword{internal}
