% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{CreateGHLabel}
\alias{CreateGHLabel}
\title{Create a GitHub label}
\usage{
CreateGHLabel(
  strLabel,
  strLabelDescription = "{qcthat}: A new label",
  strLabelColor = "#444444",
  lglVerbose = getOption("qcthat-verbose", FALSE),
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strLabel}{(\verb{length-1 character}) The name of the label to create.}

\item{strLabelDescription}{(\verb{length-1 character}) The description for the
label.}

\item{strLabelColor}{(\verb{length-1 character}) The hex color code for the
label (e.g., \code{"#444444"}).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The raw label object as returned by \code{\link[gh:gh]{gh::gh()}} (invisibly).
}
\description{
Create a GitHub label
}
\keyword{internal}
