% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCommits.R
\name{EnframeHunk}
\alias{EnframeHunk}
\title{Turn a hunk list into a tibble}
\usage{
EnframeHunk(lHunk, strFilePath, envCall = rlang::caller_env())
}
\arguments{
\item{lHunk}{(\code{list}) A single element of \code{lBlameRaw$hunks} as returned by
\code{\link[=BlameFileRaw]{BlameFileRaw()}}.}

\item{strFilePath}{(\verb{length-1 character}) A file path.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{File}: Path to the file, relative to the package root.
\item \code{Line}: Line number (integer).
\item \code{Commits}: List column where each element contains the commit SHA
(character) that last modified that line.
}
}
\description{
Turn a hunk list into a tibble
}
\keyword{internal}
