% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{ExpectUserAccepts}
\alias{ExpectUserAccepts}
\title{Does a user accept the feature?}
\usage{
ExpectUserAccepts(
  strDescription,
  intIssue,
  chrInstructions = character(),
  chrChecks = character(),
  lglReportFailure = IsCheckingUAT(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strDescription}{(\verb{length-1 character}) A brief description of a user
expectation.}

\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{chrInstructions}{(\code{character}) Instructions for how to review an issue.
Included in the associated issue before the checklist.}

\item{chrChecks}{(\code{character}) Items for the user to check. These will be
preceded by checkboxes in the associated issue.}

\item{lglReportFailure}{(\verb{length-1 logical}) Whether to ignore failures
(default unless a "qcthat_UAT" environment variable is "true"), or fail
(and show as a failure in testthat tests).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The input \code{chrChecks}, invisibly.
}
\description{
Create and track a sub-issue to track user acceptance that an issue is
complete.
}
