% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectUserAccepts.R
\name{ExpectUserAccepts}
\alias{ExpectUserAccepts}
\title{Does a user accept the feature?}
\usage{
ExpectUserAccepts(
  chrChecks,
  intIssue,
  chrInstructions = character(),
  strFailureMode = c("ignore", "fail"),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{chrChecks}{(\code{character}) Items for the user to check. These will be
preceded by checkboxes in the associated issue.}

\item{intIssue}{(\verb{length-1 integer}) The issue with which a check is
associated.}

\item{chrInstructions}{(\code{character}) Instructions for how to review an issue.
Included in the associated issue before the checklist.}

\item{strFailureMode}{(\verb{length-1 character}) Whether to \code{"ignore"} failures
(default) or \code{"fail"} (and show as a failure in testthat tests).}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The input \code{chrChecks}, invisibly.
}
\description{
Create and track a sub-issue to track user acceptance that an issue is
complete.
}
