% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTestsFromFiles.R
\name{ExtractTestsFromFile}
\alias{ExtractTestsFromFile}
\title{Extract test information from a single file}
\usage{
ExtractTestsFromFile(strFilePath, envCall = rlang::caller_env())
}
\arguments{
\item{strFilePath}{(\verb{length-1 character}) A file path.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:caller_env]{rlang::caller_env()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the same structure as
\code{\link[=ExtractTestsFromFiles]{ExtractTestsFromFiles()}}.
}
\description{
Extract test information from a single file
}
\keyword{internal}
