% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupGHLabels.R
\name{FetchGHLabels}
\alias{FetchGHLabels}
\title{Fetch existing GitHub labels}
\usage{
FetchGHLabels(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A character vector of existing label names in the specified GitHub
repository.
}
\description{
Fetch existing GitHub labels
}
\keyword{internal}
