% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{FetchPRActionRuns}
\alias{FetchPRActionRuns}
\title{Fetch workflow runs for a PR}
\usage{
FetchPRActionRuns(
  strPRHeadRef,
  strAction = character(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strPRHeadRef}{(\verb{length-1 character}) The branch name (head ref) of the
PR.}

\item{strAction}{(\verb{length-1 character}) Optional string to filter workflow
paths (e.g., "qcthat").}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}
}
\value{
A list of workflow run objects returned by \code{\link[=CallGHAPI]{CallGHAPI()}}.
}
\description{
Fetch workflow runs for a PR
}
\keyword{internal}
