% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchPRIssueNumbers.R
\name{FetchPRRefs}
\alias{FetchPRRefs}
\title{Fetch the source and target refs for a PR}
\usage{
FetchPRRefs(
  intPRNumber = GuessPRNumber(strPkgRoot, strOwner, strRepo, strGHToken),
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about and/or post results to.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions
appropriate to the action being performed.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A named character vector with \code{strSourceRef} and \code{strTargetRef}.
}
\description{
Fetch the source and target refs for a PR
}
\keyword{internal}
