% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCPR.R
\name{FetchPRRefs}
\alias{FetchPRRefs}
\title{Fetch the source and target refs for a PR}
\usage{
FetchPRRefs(
  intPRNumber = GuessPRNumber(".", strOwner, strRepo, strGHToken),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A named character vector with \code{strSourceRef} and \code{strTargetRef}.
}
\description{
Fetch the source and target refs for a PR
}
\keyword{internal}
