% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchRepoIssues.R
\name{FetchRawRepoIssues}
\alias{FetchRawRepoIssues}
\title{Fetch all repo issues from GitHub}
\usage{
FetchRawRepoIssues(
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A list of raw issue objects as returned by \code{\link[gh:gh]{gh::gh()}}.
}
\description{
Fetch all repo issues from GitHub
}
\keyword{internal}
