% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{FetchRefPRNumber}
\alias{FetchRefPRNumber}
\title{Fetch the pull request number for a branch or other git ref}
\usage{
FetchRefPRNumber(
  strSourceRef = GetActiveBranch(),
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
An integer pull request number, or \code{integer(0)} if no matching PR
(or more than one matching PR) is found.
}
\description{
Fetch the pull request number for a branch or other git ref
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  FetchRefPRNumber()
\dontshow{\}) # examplesIf}
}
