% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatFooter}
\alias{FormatFooter}
\title{Prepare the bottom of the qcthat_IssueTestMatrix output}
\usage{
FormatFooter(x)
}
\arguments{
\item{x}{(\code{qcthat_IssueTestMatrix}) The issue-test matrix to format.}
}
\value{
If \code{x} has rows, a character vector representing the formatted
footer. \code{NULL} otherwise.
}
\description{
Prepare the bottom of the qcthat_IssueTestMatrix output
}
\keyword{internal}
