% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTestMatrix.R
\name{FormatHeader}
\alias{FormatHeader}
\title{Prepare the top of the qcthat_IssueTestMatrix output}
\usage{
FormatHeader(x)
}
\arguments{
\item{x}{(\code{qcthat_IssueTestMatrix}) The issue-test matrix to format.}
}
\value{
A character vector representing the formatted header.
}
\description{
Prepare the top of the qcthat_IssueTestMatrix output
}
\keyword{internal}
