% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{GetGHAPRNumber}
\alias{GetGHAPRNumber}
\title{Get the PR number for a GitHub Action}
\usage{
GetGHAPRNumber()
}
\value{
An integer pull request number, or \code{NULL} if not running in a
PR-triggered action (as indicated by the environment variables
\code{"GITHUB_EVENT_NAME"} and \code{"GITHUB_REF_NAME"}).
}
\description{
Get the PR number for a GitHub Action
}
\keyword{internal}
