% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessPRNumber.R
\name{GuessPRNumber}
\alias{GuessPRNumber}
\title{Guess the most relevant pull request number}
\usage{
GuessPRNumber(
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to a directory in the
package. Will be expanded using \code{\link[gert:git_repo]{gert::git_find()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
The associated pull request number as an integer, or \code{NULL} if no
pull requests are found.
}
\description{
Tries to find a pull request associated with the active branch.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  GuessPRNumber()
\dontshow{\}) # examplesIf}
}
