% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRepoIssuesToCommits.R
\name{MapIssueClosersToCommits}
\alias{MapIssueClosersToCommits}
\title{Add Commits list column to dfIssueClosers}
\usage{
MapIssueClosersToCommits(
  dfIssueClosers,
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{dfIssueClosers}{(\code{data.frame}) The \code{\link[tibble:tibble]{tibble::tibble()}} returned by
\code{\link[=FetchRepoIssueClosers]{FetchRepoIssueClosers()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\itemize{
\item \code{Issue}: Issue number (integer).
\item \code{Commits}: List column containing character vectors of commit SHAs
associated with each issue.
}
}
\description{
Add Commits list column to dfIssueClosers
}
