% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestFilesToPotentialIssues.R
\name{MapLongIssueCommits}
\alias{MapLongIssueCommits}
\title{Map issues to commits in long format}
\usage{
MapLongIssueCommits(
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per issue-commit pair, containing
columns \code{Issue} and \code{Commits}.
}
\description{
Map issues to commits in long format
}
\keyword{internal}
