% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReports.R
\name{QCCompletedIssues}
\alias{QCCompletedIssues}
\title{Generate a QC report of completed issues}
\usage{
QCCompletedIssues(
  strPkgRoot = ".",
  strOwner = gh::gh_tree_remote(strPkgRoot)[["username"]],
  strRepo = gh::gh_tree_remote(strPkgRoot)[["repo"]],
  strGHToken = gh::gh_token(),
  chrIgnoredLabels = DefaultIgnoreLabels()
)
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to issues that were closed as completed.
}
\description{
A simple wrapper around \code{\link[=QCPackage]{QCPackage()}} that filters the resulting issue-test
matrix to only include issues that were closed as "completed".
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  QCCompletedIssues()
\dontshow{\}) # examplesIf}
}
