% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeGH.R
\name{QCMergeGH}
\alias{QCMergeGH}
\title{Generate a QC report of issues associated with a GitHub merge}
\usage{
QCMergeGH(
  strSourceRef = GetActiveBranch(strPkgRoot),
  strTargetRef = GetDefaultBranch(strPkgRoot),
  strPkgRoot = ".",
  strOwner = gh::gh_tree_remote(strPkgRoot)[["username"]],
  strRepo = gh::gh_tree_remote(strPkgRoot)[["repo"]],
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to issues that are associated with pull requests that will be
merged when \code{strSourceRef} is merged into \code{strTargetRef}.
}
\description{
Finds all commits in \code{strSourceRef} that are not in \code{strTargetRef}, finds all
pull requests associated with those commits, finds all issues associated with
those pull requests, and generates a QC report for those issues.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # This will only make sense if you are working in a git repository and have
  # an active branch that is different from the default branch.
  QCMergeGH()
\dontshow{\}) # examplesIf}
}
