% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMergeLocal.R
\name{QCMergeLocal}
\alias{QCMergeLocal}
\title{Generate a QC report of issues probably related to changes between local git
refs}
\usage{
QCMergeLocal(
  strSourceRef = GetActiveBranch(strPkgRoot),
  strTargetRef = GetDefaultBranch(strPkgRoot),
  strPkgRoot = ".",
  chrKeywords = c("close", "closes", "closed", "fix", "fixes", "fixed", "resolve",
    "resolves", "resolved"),
  strOwner = gh::gh_tree_remote(strPkgRoot)[["username"]],
  strRepo = gh::gh_tree_remote(strPkgRoot)[["repo"]],
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels(),
  intMaxCommits = 100000L
)
}
\arguments{
\item{strSourceRef}{(\verb{length-1 character}) Name of the git reference that
contains changes. Defaults to the active branch in this repository.}

\item{strTargetRef}{(\verb{length-1 character}) Name of the git reference that
will be merged into. Defaults to the default branch of this repository.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{chrKeywords}{(\code{character}) Keywords to search for just before issue
numbers in commit messages. Defaults to the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub issue-linking keywords}}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}

\item{intMaxCommits}{(\verb{length-1 integer}) The maximum number of commits to
return from git logs. Leaving this at the default should almost always be
fine, but you can reduce the number if your repository has a long commit
history and this function is slow.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to issues that will be closed by merging \code{strSourceRef} into
\code{strTargetRef}, using the \href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword}{GitHub keywords for linking issues to pull requests}.
}
\description{
Find issues associated with merging a source ref into a target ref based
purely on GitHub commit keywords, and generate a report about their test
status. This report is useful while you're developing a feature on a branch
other than the default branch, but you have not yet created a pull request on
GitHub.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # This will only make sense if you are working in a git repository and have
  # an active branch that is different from the default branch.
  QCMergeLocal()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=QCMergeGH]{QCMergeGH()}} to use the GitHub API to find more formal, concrete
connections between issues and the commits that closed them, and \code{\link[=QCPR]{QCPR()}}
for a wrapper around \code{\link[=QCMergeGH]{QCMergeGH()}} that specifically looks at the commits
associated with a specified pull request.
}
