% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReports.R
\name{QCPackage}
\alias{QCPackage}
\title{Generate a QC report for a package}
\usage{
QCPackage(
  strPkgRoot = ".",
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{envCall}{(\code{environment}) The environment to use for error reporting.
Typically set to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by
\code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}}.
}
\description{
A QC report combines information about GitHub issues associated with a
package (see \code{\link[=FetchRepoIssues]{FetchRepoIssues()}}) and testthat test results for the package
(see \code{\link[=CompileTestResults]{CompileTestResults()}}) using \code{\link[=CompileIssueTestMatrix]{CompileIssueTestMatrix()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  QCPackage()
\dontshow{\}) # examplesIf}
}
