% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCPR.R
\name{QCPR}
\alias{QCPR}
\title{Generate a QC report of issues associated with a GitHub pull request}
\usage{
QCPR(
  intPRNumber = GuessPRNumber(strPkgRoot, strOwner, strRepo, strGHToken),
  strPkgRoot = ".",
  strOwner = gh::gh_tree_remote(strPkgRoot)[["username"]],
  strRepo = gh::gh_tree_remote(strPkgRoot)[["repo"]],
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels()
)
}
\arguments{
\item{intPRNumber}{(\verb{length-1 integer}) The number of the pull request to
fetch information about.}

\item{strPkgRoot}{(\verb{length-1 character}) The path to the root directory of
the package. Will be expanded using \code{\link[pkgload:packages]{pkgload::pkg_path()}}.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}

\item{lglWarn}{(\verb{length-1 logical}) Whether to warn when an extra value is
included in the filter (but the report still returns results). Defaults to
\code{TRUE}.}

\item{chrIgnoredLabels}{(\code{character}) GitHub labels to ignore, such as
\code{"qcthat-nocov"}.}
}
\value{
A \code{qcthat_IssueTestMatrix} object as returned by \code{\link[=QCPackage]{QCPackage()}},
filtered to issues that will be closed by merging the pull request.
}
\description{
Find issues associated with a GitHub pull request, whether they were added
via keywords, using the pull request sidebar, or using the issue sidebar. See
\href{https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue}{GitHub Docs: Link a pull request to an issue}
for details on how issues can become associated with a pull request.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

  # You must have at least one pull request open in the GitHub repository
  # associated with the current git repository for this to return any
  # results.
  QCPR()
\dontshow{\}) # examplesIf}
}
