% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcthat-package.R
\docType{package}
\name{qcthat-package}
\alias{qcthat}
\alias{qcthat-package}
\title{qcthat: Lightweight quality control framework for R packages used in Clinical Trials}
\description{
Lightweight quality control framework for R packages used in Clinical Trials. Minimal dependencies, flexible, and customizable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilead-biostats.github.io/qcthat/}
  \item \url{https://github.com/Gilead-BioStats/qcthat}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/qcthat/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Jon Harmon \email{jon.harmon@atorusresearch.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
  \item Natalia Andriychuk \email{Natalia.Andriychuk@gilead.com}
  \item Matt Roumaya \email{Matt.Roumaya@atorusresearch.com} (\href{https://orcid.org/0000-0002-8009-2771}{ORCID})
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
