% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtils.R
\name{qcthatPath}
\alias{qcthatPath}
\title{Find a path within qcthat}
\usage{
qcthatPath(chrPath, strExtension)
}
\arguments{
\item{chrPath}{(\code{character}) Components of a path}

\item{strExtension}{(\verb{length-1 character}) The file extension to use for the
target file. If the target path already includes an extension, it will be
replaced with this value. If the value is already correct, this won't have
any effect.}
}
\value{
The full path to the file within the \code{qcthat} package.
}
\description{
Exists primarily to make it easier to mock this for testing.
}
\keyword{internal}
