% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTestsToCoveredLines.R
\name{SetupTempSnapDir}
\alias{SetupTempSnapDir}
\title{Create a temp snapshot directory with a copy of the relevant snapshot file}
\usage{
SetupTempSnapDir(strTestFile, strSnapDir)
}
\arguments{
\item{strTestFile}{(\verb{length-1 character}) Original test file path.}

\item{strSnapDir}{(\verb{length-1 character}) Path to the real \verb{_snaps} directory.}
}
\value{
Path to the temp snapshot directory.
}
\description{
Create a temp snapshot directory with a copy of the relevant snapshot file
}
\keyword{internal}
