% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchUAIssue.R
\name{TitleUAIssue}
\alias{TitleUAIssue}
\title{Generate a title for a UAT issue}
\usage{
TitleUAIssue(chrChecks)
}
\arguments{
\item{chrChecks}{(\code{character}) Items for the user to check. These will be
preceded by checkboxes in the associated issue.}
}
\value{
A string title for the UAT issue.
}
\description{
Generate a title for a UAT issue
}
\keyword{internal}
