% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriggerUAT.R
\name{TriggerUAT}
\alias{TriggerUAT}
\title{Trigger the UAT cycle for closed issues}
\usage{
TriggerUAT(
  intClosedIssue = GuessIssueNumber(),
  strOwner = GetGHOwner(),
  strRepo = GetGHRepo(),
  strGHToken = gh::gh_token()
)
}
\arguments{
\item{intClosedIssue}{(\verb{length-1 integer}) A closed UAT issue number.}

\item{strOwner}{(\verb{length-1 character}) GitHub username or organization name.}

\item{strRepo}{(\verb{length-1 character}) GitHub repository name.}

\item{strGHToken}{(\verb{length-1 character}) GitHub token with permissions to
read issues.}
}
\value{
A \code{data.frame} of PRs that referenced the issue (invisibly). This
function is called for its side effects.
}
\description{
Identify pull requests that mention a closed issue, and potentially
re-trigger the UAT cycle. If all issues referenced in the UAT comment of a
linked PR are closed (and it is not already running the QCPR workflow), rerun
the QCPR workflow (\code{qcthat-pr_issues.yaml}) if available.
}
